refactor: extract GlanceSidebar from DesktopLayout
Extract the duplicated glance tab content from both the tablet side panel
and the desktop side panel into a single GlanceSidebar component. Uses a
variant prop ("tablet" | "desktop") to handle the minor CSS differences
(hover vs active states, container styling).

DesktopLayout drops from ~4820 lines to ~2905 lines.
