+++ src/components/DesktopLayout.jsx
@@ -11,17 +11,15 @@ import {
 import { isNativeAndroid, nativeUpdateEvent } from '../native.js';
 import { renderTitle, renderTitleWithoutTags, getLinkUrl, hasNotesOrSubtasks, isLinkOnlyTask, hasOnlySubtasks, isObsidianNoteOnlyTask } from '../utils/textFormatting.jsx';
 import { dateToString, extractTags, extractWikilinks, formatDate, formatDateRange, formatDeadlineDate, formatShortDate, stripWikilinks } from '../utils/taskUtils.js';
-import { HABIT_COLORS, HABIT_ICONS } from '../constants/habits.js';
-import { HabitRing, MiniHabitRing } from './HabitRing.jsx';
-import GettingStartedChecklist from './GettingStartedChecklist.jsx';
+import { MiniHabitRing } from './HabitRing.jsx';
 import NotesSubtasksPanel from './NotesSubtasksPanel.jsx';
 import SuggestionAutocomplete from './SuggestionAutocomplete.jsx';
 import DailyNotesModal from './DailyNotesModal.jsx';
-import FrameNudgeCard from './FrameNudgeCard.jsx';
 import DeadlinePickerPopover from './DeadlinePickerPopover.jsx';
 import DesktopHeader from './DesktopHeader.jsx';
 import InboxFilterPopover from './InboxFilterPopover.jsx';
 import InboxArchivedBar from './InboxArchivedBar.jsx';
+import GlanceSidebar from './GlanceSidebar.jsx';
 import { useDayPlannerCtx } from '../context/DayPlannerContext.jsx';
 
 const DesktopLayout = () => {
@@ -666,967 +664,7 @@ const DesktopLayout = () => {
                 {/* Glance section — shown when glance tab active */}
                 {tabletActiveTab === 'glance' && (
                   <div className="p-4">
-                    <div className="space-y-4">
-                      {/* Search bar + filter */}
-                      <div className="flex items-center gap-2">
-                        <button
-                          onClick={() => { setShowSpotlight(true); playUISound('spotlight'); }}
-                          className={`flex-1 flex items-center gap-2 px-3 py-2.5 rounded-lg ${darkMode ? 'bg-white/10 text-gray-400' : 'bg-black/5 text-stone-400'} transition-colors active:opacity-70`}
-                        >
-                          <Search size={16} />
-                          <span className="text-sm">Search tasks...</span>
-                        </button>
-                        {allTags.length > 0 && (
-                          <div className="flex-shrink-0 self-stretch flex items-center">
-                            <button
-                              ref={tagFilterBtnRef}
-                              onClick={() => setShowMobileTagFilter(v => !v)}
-                              className={`px-2.5 h-full flex items-center rounded-lg transition-colors ${
-                                !allTags.every(tag => selectedTags.includes(tag))
-                                  ? 'bg-blue-500 text-white'
-                                  : darkMode ? 'bg-white/10 text-gray-400' : 'bg-black/5 text-stone-400'
-                              } active:opacity-80`}
-                            >
-                              <Filter size={16} />
-                            </button>
-                            {/* Tag filter popover — matching desktop */}
-                            {showMobileTagFilter && (() => {
-                              const rect = tagFilterBtnRef.current?.getBoundingClientRect();
-                              return (
-                              <>
-                                <div className="fixed inset-0 z-40" onClick={() => setShowMobileTagFilter(false)} />
-                                <div
-                                  className={`fixed z-50 ${cardBg} border ${borderClass} rounded-xl shadow-xl`}
-                                  style={{ width: '280px', top: rect ? rect.bottom + 4 : 0, left: rect ? Math.max(8, rect.right - 280) : 0 }}
-                                  onClick={(e) => e.stopPropagation()}
-                                >
-                                  <div className="flex items-center justify-between px-3 py-2 border-b" style={{ borderColor: 'inherit' }}>
-                                    <div className="flex items-center gap-1.5">
-                                      <Filter size={14} className={textSecondary} />
-                                      <span className={`text-sm font-semibold ${textPrimary}`}>Filter by Tag</span>
-                                    </div>
-                                    <div className="flex items-center gap-2">
-                                      {allTags.every(tag => selectedTags.includes(tag)) ? (
-                                        <button onClick={clearTagFilter} className="text-xs text-blue-500 hover:text-blue-600 font-medium">Clear</button>
-                                      ) : (
-                                        <button onClick={selectAllTags} className="text-xs text-blue-500 hover:text-blue-600 font-medium">Select All</button>
-                                      )}
-                                    </div>
-                                  </div>
-                                  <div className="py-1 max-h-[300px] overflow-y-auto">
-                                    {allTags.map(tag => {
-                                      const visibleDateStrs = new Set(visibleDates.map(d => dateToString(d)));
-                                      const regularCount = tasks.filter(t => !t.imported && visibleDateStrs.has(t.date) && extractTags(t.title).includes(tag)).length;
-                                      const recurringCount = expandedRecurringTasks.filter(t => visibleDateStrs.has(t.date) && extractTags(t.title).includes(tag)).length;
-                                      const tagCount = regularCount + recurringCount;
-                                      return (
-                                        <button
-                                          key={tag}
-                                          onClick={() => toggleTag(tag)}
-                                          className={`w-full flex items-center gap-2 px-3 py-2 transition-colors ${
-                                            tagCount === 0 ? 'opacity-40' : ''
-                                          } ${
-                                            selectedTags.includes(tag)
-                                              ? darkMode ? 'bg-blue-500/20' : 'bg-blue-50'
-                                              : darkMode ? 'hover:bg-white/5' : 'hover:bg-stone-50'
-                                          }`}
-                                        >
-                                          <div className={`w-4 h-4 rounded flex items-center justify-center flex-shrink-0 border transition-colors ${
-                                            selectedTags.includes(tag) ? 'bg-blue-500 border-blue-500' : darkMode ? 'border-gray-600' : 'border-stone-300'
-                                          }`}>
-                                            {selectedTags.includes(tag) && <Check size={12} className="text-white" />}
-                                          </div>
-                                          <Hash size={12} className={textSecondary} />
-                                          <span className={`flex-1 text-left text-sm ${textPrimary}`}>{tag}</span>
-                                          {tagCount > 0 && <span className={`text-xs ${textSecondary} tabular-nums`}>{tagCount}</span>}
-                                        </button>
-                                      );
-                                    })}
-                                  </div>
-                                </div>
-                              </>
-                              );
-                            })()}
-                          </div>
-                        )}
-                        {aiConfig.enabled && aiConfig.features.voiceTaskInput && (
-                          <button
-                            onClick={() => setShowVoiceInput(true)}
-                            className={`flex-shrink-0 self-stretch flex items-center px-2.5 rounded-lg transition-colors ${darkMode ? 'bg-white/10 text-purple-400' : 'bg-black/5 text-purple-600'} active:opacity-80`}
-                            title="Voice Task Input"
-                          >
-                            <Mic size={16} />
-                          </button>
-                        )}
-                      </div>
-
-                      {/* Morning dayGLANCE — AI morning summary card */}
-                      {aiConfig.enabled && aiConfig.features.morningSummary && !morningGlanceDismissed && (
-                        (morningGlanceText || morningGlanceLoading || morningGlanceError) ? (
-                        <div className={`rounded-lg border p-3 ${darkMode ? 'border-amber-800/50 bg-amber-900/20' : 'border-amber-200 bg-amber-50'}`}>
-                          <div className="flex items-start justify-between gap-2">
-                            <div className="flex items-center gap-2 flex-shrink-0">
-                              <Sun size={16} className="text-amber-500" />
-                              <span className={`text-xs font-semibold uppercase tracking-wider ${darkMode ? 'text-amber-300' : 'text-amber-700'}`}>Morning Briefing</span>
-                            </div>
-                            <div className="flex items-center gap-1 flex-shrink-0">
-                              <button
-                                onClick={generateMorningSummary}
-                                className={`p-1 rounded transition-colors ${darkMode ? 'hover:bg-white/10' : 'hover:bg-black/5'}`}
-                                title="Regenerate"
-                              >
-                                <RefreshCw size={12} className={`${morningGlanceLoading ? 'animate-spin' : ''} ${textSecondary}`} />
-                              </button>
-                              <button
-                                onClick={dismissMorningGlance}
-                                className={`p-1 rounded transition-colors ${darkMode ? 'hover:bg-white/10' : 'hover:bg-black/5'}`}
-                                title="Dismiss for today"
-                              >
-                                <X size={12} className={textSecondary} />
-                              </button>
-                            </div>
-                          </div>
-                          <div className="mt-2">
-                            {morningGlanceLoading && (
-                              <div className="flex items-center gap-2">
-                                <Loader size={14} className={`animate-spin ${textSecondary}`} />
-                                <span className={`text-xs ${textSecondary}`}>Generating your briefing...</span>
-                              </div>
-                            )}
-                            {morningGlanceError && (
-                              <p className={`text-xs ${darkMode ? 'text-red-400' : 'text-red-600'}`}>{morningGlanceError}</p>
-                            )}
-                            {morningGlanceText && !morningGlanceLoading && (
-                              <p className={`text-sm leading-relaxed ${textPrimary}`}>{morningGlanceText}</p>
-                            )}
-                            {morningGlanceText && !morningGlanceLoading && aiConfig?.enabled && aiConfig.features?.smartScheduling && gtdFrames.filter(f => f.enabled).length > 0 && unscheduledTasks.filter(t => !t.completed && !t.isExample).length > 0 && (
-                              <button
-                                onClick={() => { setShowFramesModal(true); setFramesModalTab('schedule'); setEditingFrame(null); }}
-                                className={`mt-2 flex items-center gap-1.5 text-xs font-medium transition-colors ${darkMode ? 'text-purple-400 hover:text-purple-300' : 'text-purple-600 hover:text-purple-700'}`}
-                              >
-                                <BrainCircuit size={12} />
-                                Schedule inbox items?
-                              </button>
-                            )}
-                          </div>
-                        </div>
-                        ) : (
-                        <div
-                          className={`flex items-center gap-2 rounded-lg border px-3 py-2 cursor-pointer transition-colors ${darkMode ? 'border-amber-800/50 bg-amber-900/20 hover:bg-amber-900/30' : 'border-amber-200 bg-amber-50 hover:bg-amber-100'}`}
-                          onClick={generateMorningSummary}
-                        >
-                          <Sun size={14} className="text-amber-500 flex-shrink-0" />
-                          <span className={`text-sm ${darkMode ? 'text-amber-300' : 'text-amber-700'}`}>Click here to see AI briefing</span>
-                          <button
-                            onClick={(e) => { e.stopPropagation(); dismissMorningGlance(); }}
-                            className={`ml-auto p-0.5 rounded flex-shrink-0 transition-colors ${darkMode ? 'hover:bg-white/10' : 'hover:bg-black/10'}`}
-                            title="Dismiss for today"
-                          >
-                            <X size={12} className={textSecondary} />
-                          </button>
-                        </div>
-                        )
-                      )}
-
-                      {/* Evening Reflection — AI end-of-day card */}
-                      {aiConfig.enabled && aiConfig.features.eveningReflection && !eveningGlanceDismissed && currentTime.getHours() >= 19 && (
-                        (eveningGlanceText || eveningGlanceLoading || eveningGlanceError) ? (
-                        <div className={`rounded-lg border p-3 ${darkMode ? 'border-indigo-800/50 bg-indigo-900/20' : 'border-indigo-200 bg-indigo-50'}`}>
-                          <div className="flex items-start justify-between gap-2">
-                            <div className="flex items-center gap-2 flex-shrink-0">
-                              <Moon size={16} className="text-indigo-400" />
-                              <span className={`text-xs font-semibold uppercase tracking-wider ${darkMode ? 'text-indigo-300' : 'text-indigo-700'}`}>Evening Reflection</span>
-                            </div>
-                            <div className="flex items-center gap-1 flex-shrink-0">
-                              <button onClick={generateEveningReflection} className={`p-1 rounded transition-colors ${darkMode ? 'hover:bg-white/10' : 'hover:bg-black/5'}`} title="Regenerate">
-                                <RefreshCw size={12} className={`${eveningGlanceLoading ? 'animate-spin' : ''} ${textSecondary}`} />
-                              </button>
-                              <button onClick={dismissEveningGlance} className={`p-1 rounded transition-colors ${darkMode ? 'hover:bg-white/10' : 'hover:bg-black/5'}`} title="Dismiss for today">
-                                <X size={12} className={textSecondary} />
-                              </button>
-                            </div>
-                          </div>
-                          <div className="mt-2">
-                            {eveningGlanceLoading && (
-                              <div className="flex items-center gap-2">
-                                <Loader size={14} className={`animate-spin ${textSecondary}`} />
-                                <span className={`text-xs ${textSecondary}`}>Reflecting on your day...</span>
-                              </div>
-                            )}
-                            {eveningGlanceError && <p className={`text-xs ${darkMode ? 'text-red-400' : 'text-red-600'}`}>{eveningGlanceError}</p>}
-                            {eveningGlanceText && !eveningGlanceLoading && <p className={`text-sm leading-relaxed ${textPrimary}`}>{eveningGlanceText}</p>}
-                            {eveningGlanceText && !eveningGlanceLoading && incompleteTodayTasks.length > 0 && gtdFrames.filter(f => f.enabled).length > 0 && aiConfig.features?.aiReschedule && (
-                              <button
-                                onClick={() => { setShowRescheduleModal(true); setRescheduleResults(null); setRescheduleError(''); }}
-                                className={`mt-2 flex items-center gap-1.5 text-xs font-medium transition-colors ${darkMode ? 'text-orange-400 hover:text-orange-300' : 'text-orange-600 hover:text-orange-700'}`}
-                              >
-                                <CalendarDays size={12} />
-                                Reschedule {incompleteTodayTasks.length} incomplete task{incompleteTodayTasks.length !== 1 ? 's' : ''} →
-                              </button>
-                            )}
-                          </div>
-                        </div>
-                        ) : (
-                        <div
-                          className={`flex items-center gap-2 rounded-lg border px-3 py-2 cursor-pointer transition-colors ${darkMode ? 'border-indigo-800/50 bg-indigo-900/20 hover:bg-indigo-900/30' : 'border-indigo-200 bg-indigo-50 hover:bg-indigo-100'}`}
-                          onClick={generateEveningReflection}
-                        >
-                          <Moon size={14} className="text-indigo-400 flex-shrink-0" />
-                          <span className={`text-sm ${darkMode ? 'text-indigo-300' : 'text-indigo-700'}`}>Click here for evening reflection</span>
-                          <button onClick={(e) => { e.stopPropagation(); dismissEveningGlance(); }} className={`ml-auto p-0.5 rounded flex-shrink-0 transition-colors ${darkMode ? 'hover:bg-white/10' : 'hover:bg-black/10'}`} title="Dismiss for today">
-                            <X size={12} className={textSecondary} />
-                          </button>
-                        </div>
-                        )
-                      )}
-
-                      {/* Frame Nudge card (tablet) */}
-                      {aiConfig.enabled && aiConfig.features?.frameNudge && activeFrameForNudge && activeFrameForNudge.minutesRemaining > 30 && agendaNowMarker.gapMinutes >= 15 && frameNudgeDismissedKey !== activeFrameNudgeKey && (frameNudgeSuggestion || frameNudgeLoading || frameNudgeError) && (
-                        <FrameNudgeCard
-                          suggestion={frameNudgeSuggestion}
-                          loading={frameNudgeLoading}
-                          error={frameNudgeError}
-                          activeFrame={activeFrameForNudge}
-                          darkMode={darkMode}
-                          textPrimary={textPrimary}
-                          textSecondary={textSecondary}
-                          onRefresh={generateFrameNudge}
-                          onDismiss={() => setFrameNudgeDismissedKey(activeFrameNudgeKey)}
-                          onStartTask={scheduleTaskAtNextSlot}
-                        />
-                      )}
-
-                      {/* Habit rings row */}
-                      {habitsEnabled && activeHabits.length > 0 && (
-                        <div className="relative">
-                          <div className="flex items-start gap-1 justify-center">
-                            {activeHabits.slice(0, 5).map((habit, habitIdx) => (
-                              <div key={habit.id} className="relative">
-                                <HabitRing
-                                  size={44}
-                                  habit={habit}
-                                  count={getTodayHabitCount(habit.id)}
-                                  darkMode={darkMode}
-                                  onClick={() => incrementHabit(habit.id)}
-                                  onContextMenu={(e) => { e.preventDefault(); setHabitLongPressId(prev => prev === habit.id ? null : habit.id); setHabitEditingCountId(null); }}
-                                  onMouseDown={() => { if (habitLongPressTimer.current) clearTimeout(habitLongPressTimer.current); habitLongPressTimer.current = setTimeout(() => { setHabitLongPressId(prev => prev === habit.id ? null : habit.id); setHabitEditingCountId(null); }, 500); }}
-                                  onMouseUp={() => { if (habitLongPressTimer.current) clearTimeout(habitLongPressTimer.current); }}
-                                  onMouseLeave={() => { if (habitLongPressTimer.current) clearTimeout(habitLongPressTimer.current); }}
-                                  onTouchStart={() => { if (habitLongPressTimer.current) clearTimeout(habitLongPressTimer.current); habitLongPressTimer.current = setTimeout(() => { setHabitLongPressId(prev => prev === habit.id ? null : habit.id); setHabitEditingCountId(null); }, 500); }}
-                                  onTouchEnd={() => { if (habitLongPressTimer.current) clearTimeout(habitLongPressTimer.current); }}
-                                />
-                                {habitLongPressId === habit.id && (
-                                  <>
-                                    <div className="fixed inset-0 z-40" onClick={() => { setHabitLongPressId(null); setHabitEditingCountId(null); }} />
-                                    <div className={`absolute top-full mt-1 z-50 ${darkMode ? 'bg-gray-800 border-gray-700' : 'bg-white border-stone-200'} border rounded-xl shadow-xl p-3 min-w-[140px] ${habitIdx === 0 ? 'left-0' : habitIdx === Math.min(activeHabits.length, 5) - 1 ? 'right-0' : 'left-1/2 -translate-x-1/2'}`}>
-                                      <div className={`text-xs font-semibold mb-2 text-center ${darkMode ? 'text-gray-300' : 'text-stone-700'}`}>{habit.name}</div>
-                                      <div className="flex items-center justify-center gap-3">
-                                        <button onClick={() => { setHabitCount(habit.id, getTodayHabitCount(habit.id) - 1); }} className={`w-8 h-8 rounded-full flex items-center justify-center ${darkMode ? 'bg-gray-700 text-gray-300 active:bg-gray-600' : 'bg-stone-100 text-stone-600 active:bg-stone-200'}`}><Minus size={16} /></button>
-                                        {habitEditingCountId === habit.id ? (
-                                    <input
-                                      type="number"
-                                      autoFocus
-                                      defaultValue={getTodayHabitCount(habit.id)}
-                                      onBlur={(e) => { setHabitCount(habit.id, parseInt(e.target.value) || 0); setHabitEditingCountId(null); }}
-                                      onKeyDown={(e) => { if (e.key === 'Enter') e.target.blur(); }}
-                                      onClick={(e) => e.stopPropagation()}
-                                      className={`w-16 text-lg font-bold text-center rounded-lg border ${darkMode ? 'bg-gray-700 text-white border-gray-600' : 'bg-stone-50 text-stone-900 border-stone-300'} outline-none focus:ring-2 focus:ring-blue-500 [appearance:textfield] [&::-webkit-inner-spin-button]:appearance-none [&::-webkit-outer-spin-button]:appearance-none`}
-                                      onFocus={(e) => e.target.select()}
-                                    />
-                                  ) : (
-                                    <span onClick={(e) => { e.stopPropagation(); setHabitEditingCountId(habit.id); }} className={`text-lg font-bold min-w-[2ch] text-center cursor-pointer hover:opacity-70 ${darkMode ? 'text-white' : 'text-stone-900'}`}>{getTodayHabitCount(habit.id)}</span>
-                                  )}
-                                        <button onClick={() => { incrementHabit(habit.id); }} className={`w-8 h-8 rounded-full flex items-center justify-center ${darkMode ? 'bg-gray-700 text-gray-300 active:bg-gray-600' : 'bg-stone-100 text-stone-600 active:bg-stone-200'}`}><Plus size={16} /></button>
-                                      </div>
-                                      <button onClick={() => { setHabitCount(habit.id, 0); setHabitLongPressId(null); setHabitEditingCountId(null); }} className="mt-2 w-full text-xs text-red-500 font-medium py-1 rounded hover:bg-red-500/10 transition-colors">Reset</button>
-                                    </div>
-                                  </>
-                                )}
-                              </div>
-                            ))}
-                            {activeHabits.length > 5 && (
-                              <div className="relative">
-                                <button
-                                  onClick={() => setHabitOverflowOpen(prev => !prev)}
-                                  className={`w-[52px] h-[44px] flex items-center justify-center rounded-lg text-xs font-bold ${darkMode ? 'bg-gray-700 text-gray-400 active:bg-gray-600' : 'bg-stone-100 text-stone-500 active:bg-stone-200'} transition-colors`}
-                                >
-                                  +{activeHabits.length - 5}
-                                </button>
-                                {habitOverflowOpen && (
-                                  <>
-                                    <div className="fixed inset-0 z-40" onClick={() => setHabitOverflowOpen(false)} />
-                                    <div className={`absolute top-full right-0 mt-1 z-50 ${darkMode ? 'bg-gray-800 border-gray-700' : 'bg-white border-stone-200'} border rounded-xl shadow-xl p-2 min-w-[180px]`}>
-                                      {activeHabits.slice(5).map(habit => {
-                                        const count = getTodayHabitCount(habit.id);
-                                        const IconComp = HABIT_ICONS[habit.icon] || Target;
-                                        const colorObj = HABIT_COLORS.find(c => c.name === habit.color) || HABIT_COLORS[0];
-                                        return (
-                                          <button
-                                            key={habit.id}
-                                            onClick={() => { incrementHabit(habit.id); }}
-                                            className={`w-full flex items-center gap-2 px-2 py-1.5 rounded-lg ${darkMode ? 'hover:bg-gray-700 active:bg-gray-600' : 'hover:bg-stone-50 active:bg-stone-100'} transition-colors`}
-                                          >
-                                            <IconComp size={16} style={{ color: colorObj.ring }} />
-                                            <span className={`text-sm flex-1 text-left ${darkMode ? 'text-gray-300' : 'text-stone-700'}`}>{habit.name}</span>
-                                            <span className={`text-xs font-semibold ${darkMode ? 'text-gray-400' : 'text-stone-500'}`}>{count}/{habit.target}</span>
-                                          </button>
-                                        );
-                                      })}
-                                    </div>
-                                  </>
-                                )}
-                              </div>
-                            )}
-                          </div>
-                        </div>
-                      )}
-                      {/* Getting Started checklist */}
-                      {showGettingStarted && <GettingStartedChecklist
-                  items={gettingStartedItems}
-                  completedCount={gettingStartedCompleteCount}
-                  darkMode={darkMode}
-                  textPrimary={textPrimary}
-                  textSecondary={textSecondary}
-                  onDismiss={() => setGettingStartedDismissed(true)}
-                  onComplete={() => {
-                    setTasks(prev => prev.filter(t => !t.isExample));
-                    setUnscheduledTasks(prev => prev.filter(t => !t.isExample));
-                    setRecycleBin(prev => prev.filter(t => !t.isExample));
-                    setRecurringTasks(prev => prev.filter(t => !t.isExample));
-                    setOnboardingComplete(true);
-                    setGettingStartedDismissed(true);
-                  }}
-                />}
-                      {/* Reschedule Tasks — shown when overdue past-day tasks exist, or incomplete today tasks after 7pm */}
-                      {aiConfig?.enabled && aiConfig.features?.aiReschedule && gtdFrames.filter(f => f.enabled).length > 0 && (() => {
-                        const _todayStr = getTodayStr();
-                        const _pastOverdue = getOverdueTasks().filter(t => t._overdueType === 'scheduled' ? t.date < _todayStr : true);
-                        if (!(_pastOverdue.length > 0 || (incompleteTodayTasks.length > 0 && currentTime.getHours() >= 19))) return null;
-                        return (
-                          <button
-                            onClick={() => { setShowRescheduleModal(true); setRescheduleResults(null); setRescheduleError(''); }}
-                            className={`w-full mb-3 flex items-center justify-center gap-2 py-2.5 px-4 rounded-xl text-sm font-semibold transition-colors ${darkMode ? 'bg-violet-600/20 hover:bg-violet-600/30 text-violet-300 border border-violet-500/30' : 'bg-violet-50 hover:bg-violet-100 text-violet-700 border border-violet-200'}`}
-                          >
-                            <Sparkles size={15} />
-                            Reschedule Tasks
-                          </button>
-                        );
-                      })()}
-                      {/* Overdue tasks from past days */}
-                      {(() => {
-                        const todayStr = getTodayStr();
-                        const pastOverdue = getOverdueTasks().filter(t => {
-                          if (t._overdueType === 'scheduled') return t.date < todayStr;
-                          return true;
-                        });
-                        if (pastOverdue.length === 0) return null;
-                        return (
-                          <div className={`rounded-lg border ${darkMode ? 'border-orange-500/40 bg-orange-500/10' : 'border-orange-400/50 bg-orange-50'} overflow-hidden`}>
-                            <button
-                              onClick={() => toggleSection('overdue')}
-                              className="w-full flex items-center justify-between px-3 py-2.5"
-                            >
-                              <div className="flex items-center gap-2">
-                                <AlertTriangle size={15} className="text-orange-500" />
-                                <span className="text-sm font-semibold text-orange-500">Overdue</span>
-                                <span className={`text-xs px-1.5 py-0.5 rounded-full ${darkMode ? 'bg-orange-500/30 text-orange-300' : 'bg-orange-200 text-orange-700'}`}>
-                                  {pastOverdue.length}
-                                </span>
-                              </div>
-                              {minimizedSections.overdue ? <ChevronDown size={16} className="text-orange-500" /> : <ChevronUp size={16} className="text-orange-500" />}
-                            </button>
-                            {!minimizedSections.overdue && (
-                              <div className="px-3 pb-2.5 space-y-1">
-                                {pastOverdue.map(task => (
-                                  <div
-                                    key={`tablet-overdue-${task.id}`}
-                                    className={`flex items-center gap-2.5 py-2 px-2 rounded-lg ${darkMode ? 'bg-white/5' : 'bg-white/80'}`}
-                                  >
-                                    <button
-                                      onClick={() => toggleComplete(task.id, task._overdueType === 'deadline')}
-                                      className={`w-5 h-5 rounded flex-shrink-0 border-2 ${task.completed
-                                        ? 'border-orange-400 bg-orange-400'
-                                        : darkMode ? 'border-orange-400/60 bg-white/10' : 'border-orange-400/60 bg-white'
-                                      } flex items-center justify-center`}
-                                    >
-                                      {task.completed && <Check size={12} strokeWidth={3} className="text-white" />}
-                                    </button>
-                                    <div
-                                      className={`flex-1 min-w-0 ${task._overdueType === 'scheduled' ? 'cursor-pointer' : ''}`}
-                                      onClick={() => {
-                                        if (task._overdueType !== 'scheduled') return;
-                                        if (task.date) setSelectedDate(new Date(task.date + 'T12:00:00'));
-                                        setTimeout(() => {
-                                          const el = document.querySelector(`[data-task-id="${task.id}"]`);
-                                          if (el && calendarRef.current) {
-                                            const container = calendarRef.current;
-                                            const elTop = el.getBoundingClientRect().top - container.getBoundingClientRect().top + container.scrollTop;
-                                            const scrollTarget = Math.max(0, elTop - container.clientHeight / 2 + el.offsetHeight / 2);
-                                            container.scrollTo({ top: scrollTarget, behavior: 'smooth' });
-                                            el.classList.add('ring-2', 'ring-blue-400');
-                                            setTimeout(() => el.classList.remove('ring-2', 'ring-blue-400'), 2000);
-                                          }
-                                        }, 200);
-                                      }}
-                                    >
-                                      <div className={`text-sm font-medium truncate ${task.completed ? 'line-through opacity-50' : textPrimary}`}>
-                                        {renderTitle(task.title)}
-                                      </div>
-                                      <div className={`text-xs ${textSecondary} flex items-center gap-1 mt-0.5`}>
-                                        {task._overdueType === 'scheduled' ? (
-                                          <><CalendarDays size={10} /> {formatDeadlineDate(task.date)}</>
-                                        ) : (
-                                          <><AlertCircle size={10} /> Due: {formatDeadlineDate(task.deadline)}</>
-                                        )}
-                                      </div>
-                                    </div>
-                                    <div className="flex items-center gap-1 flex-shrink-0">
-                                      {task.isRecurring ? (
-                                        <button
-                                          onClick={() => toggleComplete(task.id, false)}
-                                          className={`p-1.5 rounded-lg ${darkMode ? 'bg-white/10 text-gray-400' : 'bg-stone-100 text-stone-500'} active:scale-95 transition-transform`}
-                                          title="Mark complete"
-                                        >
-                                          <CheckCircle size={14} />
-                                        </button>
-                                      ) : (
-                                        <button
-                                          onClick={() => {
-                                            if (task._overdueType === 'scheduled') {
-                                              pushUndo();
-                                              setTasks(prev => prev.filter(t => t.id !== task.id));
-                                              const { startTime, date, duration, _overdueType, ...rest } = task;
-                                              setUnscheduledTasks(prev => [...prev, { ...rest, priority: rest.priority || 0 }]);
-                                              playUISound('slide');
-                                              setUndoToast({ message: 'Moved to inbox', actionable: true });
-                                            } else {
-                                              clearDeadline(task.id);
-                                              playUISound('slide');
-                                              setUndoToast({ message: 'Deadline cleared', actionable: true });
-                                            }
-                                          }}
-                                          className={`p-1.5 rounded-lg ${darkMode ? 'bg-white/10 text-gray-400' : 'bg-stone-100 text-stone-500'} active:scale-95 transition-transform`}
-                                          title="Move to inbox"
-                                        >
-                                          <Inbox size={14} />
-                                        </button>
-                                      )}
-                                      <button
-                                        onClick={() => moveToRecycleBin(task.id, task._overdueType === 'deadline')}
-                                        className={`p-1.5 rounded-lg ${darkMode ? 'bg-white/10 text-gray-400' : 'bg-stone-100 text-stone-500'} active:scale-95 transition-transform`}
-                                        title="Move to Recycle Bin"
-                                      >
-                                        <Trash2 size={14} />
-                                      </button>
-                                    </div>
-                                  </div>
-                                ))}
-                              </div>
-                            )}
-                          </div>
-                        );
-                      })()}
-
-                      {/* Today's agenda — grouped by frames */}
-                      {(() => {
-                        const filteredAgenda = filterByTags(projectFilter ? todayAgenda.filter(t => t.projectId === projectFilter) : todayAgenda);
-                        const today = new Date(getTodayStr() + 'T12:00:00');
-                        const nowMinGlance = currentTime.getHours() * 60 + currentTime.getMinutes();
-                        const todayFrames = getFrameInstancesForDate(today).filter(f => timeToMinutes(f.end) > nowMinGlance);
-                        const glanceBorderColorMap = darkMode ? {
-                          'bg-indigo-200': 'rgba(165,180,252,0.4)',
-                          'bg-amber-200': 'rgba(253,230,138,0.4)',
-                          'bg-green-200': 'rgba(167,243,208,0.4)',
-                          'bg-blue-200': 'rgba(191,219,254,0.4)',
-                          'bg-rose-200': 'rgba(254,205,211,0.4)',
-                          'bg-purple-200': 'rgba(221,214,254,0.4)',
-                          'bg-teal-200': 'rgba(153,246,228,0.4)',
-                          'bg-orange-200': 'rgba(254,215,170,0.4)',
-                        } : {
-                          'bg-indigo-200': 'rgba(79,70,229,0.75)',
-                          'bg-amber-200': 'rgba(217,119,6,0.75)',
-                          'bg-green-200': 'rgba(22,163,74,0.75)',
-                          'bg-blue-200': 'rgba(37,99,235,0.75)',
-                          'bg-rose-200': 'rgba(225,29,72,0.75)',
-                          'bg-purple-200': 'rgba(147,51,234,0.75)',
-                          'bg-teal-200': 'rgba(13,148,136,0.75)',
-                          'bg-orange-200': 'rgba(234,88,12,0.75)',
-                        };
-                        const glanceColorMap = darkMode ? {
-                          'bg-indigo-200': 'rgba(165,180,252,0.08)',
-                          'bg-amber-200': 'rgba(253,230,138,0.08)',
-                          'bg-green-200': 'rgba(167,243,208,0.08)',
-                          'bg-blue-200': 'rgba(191,219,254,0.08)',
-                          'bg-rose-200': 'rgba(254,205,211,0.08)',
-                          'bg-purple-200': 'rgba(221,214,254,0.08)',
-                          'bg-teal-200': 'rgba(153,246,228,0.08)',
-                          'bg-orange-200': 'rgba(254,215,170,0.08)',
-                        } : {
-                          'bg-indigo-200': 'rgba(165,180,252,0.18)',
-                          'bg-amber-200': 'rgba(253,230,138,0.18)',
-                          'bg-green-200': 'rgba(167,243,208,0.18)',
-                          'bg-blue-200': 'rgba(191,219,254,0.18)',
-                          'bg-rose-200': 'rgba(254,205,211,0.18)',
-                          'bg-purple-200': 'rgba(221,214,254,0.18)',
-                          'bg-teal-200': 'rgba(153,246,228,0.18)',
-                          'bg-orange-200': 'rgba(254,215,170,0.18)',
-                        };
-
-                        // Classify each agenda task into a frame or "unframed"
-                        const taskFrameMap = new Map(); // taskId -> frameId
-                        for (const task of filteredAgenda) {
-                          if (task._agendaType !== 'scheduled' || !task.startTime) continue;
-                          const tStart = timeToMinutes(task.startTime);
-                          const tEnd = tStart + (task.duration || 30);
-                          for (const frame of todayFrames) {
-                            const fStart = timeToMinutes(frame.start);
-                            const fEnd = timeToMinutes(frame.end);
-                            if (tStart >= fStart && tEnd <= fEnd) {
-                              taskFrameMap.set(String(task.id), frame.frameId);
-                              break;
-                            }
-                          }
-                        }
-
-                        // Build ordered sections: non-scheduled tasks first, then chronological frame/unframed groups
-                        const nonScheduled = filteredAgenda.filter(t => t._agendaType !== 'scheduled');
-                        const scheduled = filteredAgenda.filter(t => t._agendaType === 'scheduled');
-
-                        // Group scheduled tasks into sections in time order
-                        const sections = []; // { type: 'frame' | 'unframed', frame?, tasks[] }
-                        const framedIds = new Set(taskFrameMap.values());
-                        // Sort frames by start time
-                        const sortedFrames = [...todayFrames].filter(f => framedIds.has(f.frameId) || true).sort((a, b) => timeToMinutes(a.start) - timeToMinutes(b.start));
-                        const assignedTaskIds = new Set();
-
-                        // Interleave frames and unframed tasks in time order
-                        let scheduledIdx = 0;
-                        for (const frame of sortedFrames) {
-                          const fStart = timeToMinutes(frame.start);
-                          // Add unframed tasks before this frame
-                          const beforeTasks = [];
-                          while (scheduledIdx < scheduled.length) {
-                            const t = scheduled[scheduledIdx];
-                            const tStart = timeToMinutes(t.startTime || '00:00');
-                            if (tStart < fStart && !taskFrameMap.has(String(t.id))) {
-                              beforeTasks.push(t);
-                              assignedTaskIds.add(String(t.id));
-                              scheduledIdx++;
-                            } else break;
-                          }
-                          if (beforeTasks.length > 0) sections.push({ type: 'unframed', tasks: beforeTasks });
-
-                          // Add this frame's tasks
-                          const frameTasks = scheduled.filter(t => taskFrameMap.get(String(t.id)) === frame.frameId);
-                          const availSlots = computeAvailableSlots(frame, today);
-                          const totalAvail = availSlots.reduce((sum, s) => sum + s.minutes, 0);
-                          // Hide frames with no availability and no tasks (fully blocked by calendar events)
-                          if (totalAvail > 0 || frameTasks.length > 0) {
-                            sections.push({ type: 'frame', frame, tasks: frameTasks, totalAvail });
-                          }
-                          frameTasks.forEach(t => assignedTaskIds.add(String(t.id)));
-                          // Advance scheduledIdx past frame tasks
-                          while (scheduledIdx < scheduled.length && assignedTaskIds.has(String(scheduled[scheduledIdx].id))) scheduledIdx++;
-                        }
-                        // Remaining unframed tasks after all frames
-                        const remaining = scheduled.filter(t => !assignedTaskIds.has(String(t.id)));
-                        if (remaining.length > 0) sections.push({ type: 'unframed', tasks: remaining });
-
-                        const nowMin = currentTime.getHours() * 60 + currentTime.getMinutes();
-
-                        // Compute now marker position within sections (frame-aware)
-                        let nowMarkerSectionInfo = null;
-                        let nowIsAfterAllSections = false;
-                        if (sections.length > 0 && !agendaNowMarker.insideTask) {
-                          for (let si = 0; si < sections.length; si++) {
-                            const section = sections[si];
-                            let sStart, sEnd;
-                            if (section.type === 'frame') {
-                              sStart = timeToMinutes(section.frame.start);
-                              sEnd = timeToMinutes(section.frame.end);
-                            } else {
-                              if (section.tasks.length === 0) continue;
-                              sStart = Math.min(...section.tasks.map(t => timeToMinutes(t.startTime || '00:00')));
-                              sEnd = Math.max(...section.tasks.map(t => timeToMinutes(t.startTime || '00:00') + (t.duration || 0)));
-                            }
-                            if (nowMin >= sStart && nowMin < sEnd) {
-                              let afterTaskIdx = -1;
-                              for (let ti = 0; ti < section.tasks.length; ti++) {
-                                const taskEnd = timeToMinutes(section.tasks[ti].startTime || '00:00') + (section.tasks[ti].duration || 0);
-                                if (nowMin >= taskEnd) afterTaskIdx = ti;
-                              }
-                              nowMarkerSectionInfo = { si, inSection: true, afterTaskIdx };
-                              break;
-                            } else if (nowMin < sStart) {
-                              nowMarkerSectionInfo = { si, inSection: false };
-                              break;
-                            }
-                          }
-                          if (!nowMarkerSectionInfo) {
-                            nowIsAfterAllSections = true;
-                          }
-                        }
-
-                        const renderTabletNowMarker = (key) => {
-                          const gapH = Math.floor(agendaNowMarker.gapMinutes / 60);
-                          const gapM = agendaNowMarker.gapMinutes % 60;
-                          const gapStr = gapH > 0 ? `${gapH}h${gapM > 0 ? ` ${gapM}m` : ''}` : `${gapM}m`;
-                          return (
-                            <div key={key} className="flex gap-2.5 py-2">
-                              <div className="w-1.5 rounded-full flex-shrink-0 bg-red-500" />
-                              <div className="min-w-0 flex-1">
-                                <div className="text-sm font-medium text-red-500">{formatTime(agendaNowMarker.nowTimeStr)}, {gapStr} of free time</div>
-                                {agendaNowMarker.gapMinutes < 30 ? (
-                                  <div className="text-xs italic text-red-500 mt-0.5">Get ready to be productive!</div>
-                                ) : agendaNowMarker.inboxCount > 0 ? (
-                                  <div className="flex items-center gap-1.5 mt-0.5">
-                                    <span className="text-xs italic text-red-500">Maybe tackle an inbox task?</span>
-                                    {agendaNowMarker.showNudge && aiConfig.enabled && aiConfig.features?.frameNudge && activeFrameForNudge && (
-                                      <button onClick={() => { setFrameNudgeDismissedKey(''); generateFrameNudge(); }} className="flex items-center gap-0.5 px-1.5 py-0.5 rounded text-[10px] font-medium bg-teal-500/20 hover:bg-teal-500/30 text-teal-600 dark:text-teal-400 transition-colors">
-                                        <Sparkles size={9} />AI
-                                      </button>
-                                    )}
-                                  </div>
-                                ) : null}
-                              </div>
-                            </div>
-                          );
-                        };
-
-                        const renderTaskItem = (task, keyPrefix) => {
-                          const colorClass = task.color === 'task-calendar' ? '' : task.color;
-                          let timeLabel = '';
-                          let relativeLabel = '';
-                          if (task._agendaType === 'allday') {
-                            timeLabel = 'ALL DAY';
-                          } else if (task._agendaType === 'deadline') {
-                            timeLabel = 'DUE TODAY';
-                          } else {
-                            const [h, m] = (task.startTime || '0:0').split(':').map(Number);
-                            const startMin = h * 60 + m;
-                            const endMin = startMin + (task.duration || 0);
-                            const endH = String(Math.floor(endMin / 60)).padStart(2, '0');
-                            const endM = String(endMin % 60).padStart(2, '0');
-                            timeLabel = `${formatTime(task.startTime)}\u00A0–\u00A0${formatTime(endH + ':' + endM)}`;
-                            const diff = startMin - nowMin;
-                            if (diff > 0) {
-                              relativeLabel = diff >= 60 ? `in ${Math.floor(diff / 60)}h ${diff % 60 > 0 ? `${diff % 60}m` : ''}` : `in ${diff}m`;
-                            } else if (diff === 0) {
-                              relativeLabel = 'now';
-                            } else if (nowMin < endMin && !task.completed) {
-                              relativeLabel = 'In Progress';
-                            } else if (nowMin >= endMin && !task.completed) {
-                              relativeLabel = 'Overdue';
-                            }
-                          }
-                          return (
-                            <div
-                              key={`${keyPrefix}-${task._agendaType}-${task.id}`}
-                              data-ctx-menu
-                              onContextMenu={(e) => {
-                                e.preventDefault();
-                                setTaskContextMenu({
-                                  x: e.clientX, y: e.clientY,
-                                  taskId: task.id,
-                                  isRecurring: !!task.isRecurring,
-                                  isImported: !!task.imported,
-                                  isAllDay: !!task.isAllDay || task._agendaType === 'allday',
-                                  dateStr: dateToString(new Date()),
-                                });
-                              }}
-                              className={`flex gap-2.5 py-2 ${task.completed ? 'opacity-50' : ''} cursor-pointer active:bg-white/5 rounded-lg transition-colors`}
-                              onClick={() => {
-                                const el = document.querySelector(`[data-task-id="${task.id}"]`);
-                                if (el && calendarRef.current) {
-                                  const container = calendarRef.current;
-                                  const elTop = el.getBoundingClientRect().top - container.getBoundingClientRect().top + container.scrollTop;
-                                  const scrollTarget = Math.max(0, elTop - container.clientHeight / 2 + el.offsetHeight / 2);
-                                  container.scrollTo({ top: scrollTarget, behavior: 'smooth' });
-                                  el.classList.add('ring-2', 'ring-blue-400');
-                                  setTimeout(() => el.classList.remove('ring-2', 'ring-blue-400'), 2000);
-                                }
-                              }}
-                            >
-                              <div className={`w-1.5 rounded-full flex-shrink-0 ${colorClass} ${relativeLabel === 'In Progress' ? 'animate-pulse' : ''}`} style={task.isTaskCalendar ? getTaskCalendarStyle(task, darkMode) : task.nativeCalendarColor ? { backgroundColor: task.nativeCalendarColor } : {}}></div>
-                              <div className="min-w-0 flex-1">
-                                <div className={`text-sm font-semibold ${textPrimary} ${task.completed ? 'line-through' : ''} flex items-center gap-1.5`}>
-                                  {task.isRecurring && <RefreshCw size={13} className="flex-shrink-0 opacity-60" />}
-                                  {task.importSource === 'obsidian' && <BookOpen size={13} className="flex-shrink-0 opacity-60" title="From Obsidian" />}
-                                  <span className="truncate">{renderTitle(task.title)}</span>
-                                  {isNativeAndroid() && extractWikilinks(task.title).map((note, i) => (
-                                    <button key={i} className="flex-shrink-0 text-purple-400 active:text-purple-300" onClick={(e) => { e.stopPropagation(); window.DayGlanceObsidian?.openNote(note); }} title={`Open "${note}" in Obsidian`}><NotebookPen size={13} /></button>
-                                  ))}
-                                </div>
-                                <div className={`text-sm ${textSecondary} flex items-center gap-1`}>
-                                  <span className="whitespace-nowrap">{timeLabel}{relativeLabel ? ',' : ''}</span>{relativeLabel ? <span className={relativeLabel === 'Overdue' ? 'text-orange-500 font-medium' : relativeLabel === 'In Progress' ? 'text-blue-500 font-medium' : ''}>{relativeLabel}</span> : ''}
-                                  {relativeLabel === 'In Progress' && focusModeAvailable && (
-                                    <button
-                                      onClick={(e) => { e.stopPropagation(); enterFocusMode(); }}
-                                      className="ml-1 p-1.5 rounded text-purple-500 active:text-purple-400 active:bg-purple-500/20 transition-colors"
-                                      title="Enter Focus Mode"
-                                    >
-                                      <Target size={16} className="animate-pulse" />
-                                    </button>
-                                  )}
-                                </div>
-                                {goalsProjectsEnabled && task.projectId && (() => {
-                                  const proj = projects.find(p => p.id === task.projectId);
-                                  if (!proj) return null;
-                                  return (
-                                    <button
-                                      onClick={(e) => { e.stopPropagation(); setProjectFilter(prev => prev === task.projectId ? null : task.projectId); }}
-                                      className={`inline-flex items-center text-[10px] px-1.5 py-0.5 rounded-full font-medium transition-colors ${darkMode ? 'bg-blue-900/50 text-blue-300 hover:bg-blue-800/70' : 'bg-blue-100 text-blue-700 hover:bg-blue-200'} ${projectFilter === task.projectId ? 'ring-1 ring-blue-400' : ''}`}
-                                      title={projectFilter === task.projectId ? 'Clear project filter' : `Filter: ${proj.title}`}
-                                    >
-                                      {proj.title}
-                                    </button>
-                                  );
-                                })()}
-                              </div>
-                              {(relativeLabel === 'Overdue' || (task._agendaType === 'allday' && !task.imported)) && !task.completed && (
-                                <div className="flex items-center gap-1 flex-shrink-0 mr-5">
-                                  {!task.isRecurring && (
-                                    <button
-                                      onClick={(e) => {
-                                        e.stopPropagation();
-                                        pushUndo();
-                                        setTasks(prev => prev.filter(t => t.id !== task.id));
-                                        const { startTime, date, _agendaType, ...rest } = task;
-                                        setUnscheduledTasks(prev => [...prev, { ...rest, priority: rest.priority || 0 }]);
-                                        playUISound('slide');
-                                        setUndoToast({ message: 'Moved to inbox', actionable: true });
-                                      }}
-                                      className={`p-1.5 rounded-lg ${darkMode ? 'bg-white/10 text-gray-400' : 'bg-stone-100 text-stone-500'} active:scale-95 transition-transform`}
-                                      title="Move to Inbox"
-                                    >
-                                      <Inbox size={14} />
-                                    </button>
-                                  )}
-                                  <button
-                                    onClick={(e) => { e.stopPropagation(); toggleComplete(task.id, false); }}
-                                    className={`p-1.5 rounded-lg ${darkMode ? 'bg-white/10 text-gray-400' : 'bg-stone-100 text-stone-500'} active:scale-95 transition-transform`}
-                                    title="Mark complete"
-                                  >
-                                    <CheckCircle size={14} />
-                                  </button>
-                                </div>
-                              )}
-                            </div>
-                          );
-                        };
-
-                        return (
-                        <div className="space-y-1.5">
-                          {filteredAgenda.length === 0 && (
-                            <p className={`text-sm ${textSecondary} text-center py-4`}>No tasks scheduled for today</p>
-                          )}
-                          {/* Now marker before first task (only when no frame sections handle positioning) */}
-                          {filteredAgenda.length > 0 && sections.length === 0 && !agendaNowMarker.insideTask && agendaNowMarker.insertAfterIndex < 0 && (() => {
-                            const gapH = Math.floor(agendaNowMarker.gapMinutes / 60);
-                            const gapM = agendaNowMarker.gapMinutes % 60;
-                            const gapStr = gapH > 0 ? `${gapH}h${gapM > 0 ? ` ${gapM}m` : ''}` : `${gapM}m`;
-                            return (
-                              <div key="tablet-now-marker" className="flex gap-2.5 py-2.5">
-                                <div className="w-1.5 rounded-full flex-shrink-0 bg-red-500" />
-                                <div className="min-w-0 flex-1">
-                                  <div className="text-sm font-medium text-red-500">{formatTime(agendaNowMarker.nowTimeStr)}, {gapStr} of free time</div>
-                                  {agendaNowMarker.gapMinutes < 30 ? (
-                                    <div className="text-xs italic text-red-500 mt-0.5">Get ready to be productive!</div>
-                                  ) : agendaNowMarker.inboxCount > 0 ? (
-                                    <div className="flex items-center gap-1.5 mt-0.5">
-                                      <span className="text-xs italic text-red-500">Maybe tackle an inbox task?</span>
-                                      {agendaNowMarker.showNudge && aiConfig.enabled && aiConfig.features?.frameNudge && activeFrameForNudge && (
-                                        <button onClick={() => { setFrameNudgeDismissedKey(''); generateFrameNudge(); }} className="flex items-center gap-0.5 px-1.5 py-0.5 rounded text-[10px] font-medium bg-teal-500/20 hover:bg-teal-500/30 text-teal-600 dark:text-teal-400 transition-colors">
-                                          <Sparkles size={9} />AI
-                                        </button>
-                                      )}
-                                    </div>
-                                  ) : null}
-                                </div>
-                              </div>
-                            );
-                          })()}
-                          {/* Non-scheduled items (all-day, deadlines) */}
-                          {nonScheduled.map(task => renderTaskItem(task, 'tablet-glance'))}
-                          {/* Frame-grouped and unframed sections */}
-                          {sections.map((section, si) => {
-                            const elements = [];
-                            // Now marker between sections (before this section)
-                            if (nowMarkerSectionInfo && !nowMarkerSectionInfo.inSection && nowMarkerSectionInfo.si === si) {
-                              elements.push(renderTabletNowMarker(`tablet-now-mid-${si}`));
-                            }
-                            if (section.type === 'frame') {
-                              const borderColor = glanceBorderColorMap[section.frame.color] || (darkMode ? 'rgba(165,180,252,0.4)' : 'rgba(79,70,229,0.75)');
-                              const bgColor = glanceColorMap[section.frame.color] || (darkMode ? 'rgba(165,180,252,0.08)' : 'rgba(165,180,252,0.18)');
-                              const availH = Math.floor(section.totalAvail / 60);
-                              const availM = section.totalAvail % 60;
-                              const availStr = availH > 0 ? `${availH}h${availM > 0 ? ` ${availM}m` : ''}` : `${availM}m`;
-                              const markerInThisFrame = nowMarkerSectionInfo && nowMarkerSectionInfo.inSection && nowMarkerSectionInfo.si === si;
-                              elements.push(
-                                <div
-                                  key={`tablet-frame-section-${section.frame.frameId}`}
-                                  className="rounded-md overflow-hidden"
-                                  style={{
-                                    borderLeft: `3px solid ${borderColor}`,
-                                    background: bgColor,
-                                  }}
-                                >
-                                  <div className="px-3 pt-2 pb-1">
-                                    <div className="flex items-center gap-1.5">
-                                      <LayoutGrid size={12} style={{ color: borderColor }} />
-                                      <span className="text-xs font-semibold" style={{ color: borderColor }}>{section.frame.label}</span>
-                                      <span className={`text-xs ${textSecondary}`}>{formatTime(section.frame.start)} – {formatTime(section.frame.end)}</span>
-                                    </div>
-                                    {section.totalAvail > 0 && (
-                                      <p className="mt-1">
-                                        <span className={`text-xs font-medium px-2 py-0.5 rounded-full ${darkMode ? 'bg-blue-900/40 text-blue-300' : 'bg-blue-100 text-blue-700'}`}>
-                                          {availStr} available
-                                        </span>
-                                      </p>
-                                    )}
-                                  </div>
-                                  <div className="px-2 pb-1.5">
-                                    {section.tasks.length === 0 && !markerInThisFrame ? (
-                                      <p className={`text-xs ${textSecondary} py-2 px-1 italic`}>No tasks scheduled</p>
-                                    ) : (() => {
-                                      const items = [];
-                                      if (markerInThisFrame && nowMarkerSectionInfo.afterTaskIdx < 0) {
-                                        items.push(renderTabletNowMarker(`tablet-now-frame-${si}`));
-                                      }
-                                      section.tasks.forEach((task, ti) => {
-                                        items.push(renderTaskItem(task, 'tablet-frame'));
-                                        if (markerInThisFrame && nowMarkerSectionInfo.afterTaskIdx === ti) {
-                                          items.push(renderTabletNowMarker(`tablet-now-frame-${si}-${ti}`));
-                                        }
-                                      });
-                                      return items;
-                                    })()}
-                                  </div>
-                                </div>
-                              );
-                            } else {
-                              // Unframed tasks
-                              const markerInThisSection = nowMarkerSectionInfo && nowMarkerSectionInfo.inSection && nowMarkerSectionInfo.si === si;
-                              if (markerInThisSection && nowMarkerSectionInfo.afterTaskIdx < 0) {
-                                elements.push(renderTabletNowMarker(`tablet-now-unframed-${si}`));
-                              }
-                              section.tasks.forEach((task, ti) => {
-                                elements.push(renderTaskItem(task, 'tablet-glance'));
-                                if (markerInThisSection && nowMarkerSectionInfo.afterTaskIdx === ti) {
-                                  elements.push(renderTabletNowMarker(`tablet-now-unframed-${si}-${ti}`));
-                                }
-                              });
-                            }
-                            return <React.Fragment key={`tablet-section-${si}`}>{elements}</React.Fragment>;
-                          })}
-                          {/* Now marker after all tasks/frames */}
-                          {filteredAgenda.length > 0 && !agendaNowMarker.insideTask && (sections.length > 0 ? nowIsAfterAllSections : agendaNowMarker.insertAfterIndex >= todayAgenda.length - 1) && (() => {
-                            const hr = currentTime.getHours();
-                            const barColor = hr >= 22 ? 'bg-blue-500' : hr >= 19 ? 'bg-green-500' : 'bg-yellow-500';
-                            const textColor = hr >= 22 ? 'text-blue-500' : hr >= 19 ? 'text-green-500' : 'text-yellow-600';
-                            const subtitle = hr >= 22 ? "Get some rest so you're ready for tomorrow!" : hr >= 19 ? 'Enjoy the evening!' : 'Time to relax or tackle more tasks?';
-                            return (
-                              <div key="tablet-now-marker-end" className="flex gap-2.5 py-2.5">
-                                <div className={`w-1.5 rounded-full flex-shrink-0 ${barColor}`} />
-                                <div className="min-w-0 flex-1">
-                                  <div className={`text-sm font-medium ${textColor}`}>{formatTime(agendaNowMarker.nowTimeStr)}, all done!</div>
-                                  <div className={`text-xs italic ${textColor} mt-0.5`}>{subtitle}</div>
-                                </div>
-                              </div>
-                            );
-                          })()}
-                        </div>
-                      ); })()}
-
-                      {/* GLANCEahead — tomorrow preview */}
-                      {(() => {
-                        const isDayDone = (todayAgenda.length > 0 && !agendaNowMarker.insideTask && agendaNowMarker.insertAfterIndex >= todayAgenda.length - 1) || todayAgenda.length === 0;
-                        const isEvening = currentTime.getHours() >= 19;
-                        if (!isDayDone && !isEvening) return null;
-                        const { dayLabel, taskCount, eventCount, deadlineCount, firstStartTime, committedMinutes, isEmpty } = glanceAhead;
-                        const committedH = Math.floor(committedMinutes / 60);
-                        const committedM = committedMinutes % 60;
-                        const committedStr = committedH > 0 ? `${committedH}h${committedM > 0 ? ` ${committedM}m` : ''}` : committedM > 0 ? `${committedM}m` : null;
-                        return (
-                          <div className={`mt-3 pt-3 border-t ${borderClass}`}>
-                            <div className={`text-xs font-semibold uppercase tracking-wide mb-2 ${textSecondary}`}>
-                              <span className="flex items-center gap-1.5">
-                                <span><span className="italic">GLANCE</span><span className="normal-case not-italic">ahead</span></span>
-                                <span className="font-normal normal-case">— {dayLabel}</span>
-                              </span>
-                            </div>
-                            {isEmpty ? (
-                              <p className={`text-sm ${textSecondary} italic`}>Tomorrow is wide open</p>
-                            ) : (
-                              <div className="space-y-1">
-                                {firstStartTime && (
-                                  <div className="flex items-center gap-2">
-                                    <Clock size={13} className={textSecondary} />
-                                    <span className={`text-sm ${textPrimary}`}>Day starts at <span className="font-medium">{formatTime(firstStartTime)}</span></span>
-                                  </div>
-                                )}
-                                <div className="flex flex-wrap items-center gap-x-3 gap-y-1">
-                                  {taskCount > 0 && (
-                                    <span className={`text-sm ${textPrimary} flex items-center gap-1`}><CheckSquare size={12} className={textSecondary} />{taskCount} task{taskCount !== 1 ? 's' : ''}</span>
-                                  )}
-                                  {eventCount > 0 && (
-                                    <span className={`text-sm ${textPrimary} flex items-center gap-1`}><Calendar size={12} className={textSecondary} />{eventCount} event{eventCount !== 1 ? 's' : ''}</span>
-                                  )}
-                                  {deadlineCount > 0 && (
-                                    <span className={`text-sm font-medium ${darkMode ? 'text-orange-400' : 'text-orange-600'} flex items-center gap-1`}><AlertTriangle size={12} />{deadlineCount} deadline{deadlineCount !== 1 ? 's' : ''}</span>
-                                  )}
-                                  {committedStr && (
-                                    <span className={`text-xs px-2 py-0.5 rounded-full ${darkMode ? 'bg-blue-900/40 text-blue-300' : 'bg-blue-100 text-blue-700'}`}>{committedStr} committed</span>
-                                  )}
-                                </div>
-                              </div>
-                            )}
-                          </div>
-                        );
-                      })()}
-
-                      {/* Routines row */}
-                      {routinesEnabled && todayRoutines.length > 0 && (() => {
-                        const nowMin = currentTime.getHours() * 60 + currentTime.getMinutes();
-                        const visibleRoutines = todayRoutines.filter(r => {
-                          if (String(r.id).startsWith('example-')) return false;
-                          if (!r.startTime || r.isAllDay) return true;
-                          return (timeToMinutes(r.startTime) + r.duration + 60) > nowMin;
-                        });
-                        if (visibleRoutines.length === 0) return null;
-                        return (
-                          <div className={`mt-3 pt-3 border-t ${borderClass} cursor-pointer active:opacity-70 transition-opacity`} onClick={() => openRoutinesDashboard()}>
-                            <div className={`text-xs font-semibold uppercase tracking-wide mb-2 ${textSecondary}`}>Routines</div>
-                            <div className="flex flex-wrap gap-1.5">
-                              {[...visibleRoutines].sort((a, b) => {
-                                if (a.isAllDay && !b.isAllDay) return -1;
-                                if (!a.isAllDay && b.isAllDay) return 1;
-                                if (a.startTime && b.startTime) return timeToMinutes(a.startTime) - timeToMinutes(b.startTime);
-                                return 0;
-                              }).map(r => {
-                                let timeLabel = '';
-                                if (!r.isAllDay && r.startTime) {
-                                  if (use24HourClock) {
-                                    timeLabel = r.startTime;
-                                  } else {
-                                    const [h, m] = r.startTime.split(':').map(Number);
-                                    const hour12 = h === 0 ? 12 : h > 12 ? h - 12 : h;
-                                    const ampm = h < 12 ? 'a' : 'p';
-                                    timeLabel = m === 0 ? `${hour12}${ampm}` : `${hour12}:${String(m).padStart(2, '0')}${ampm}`;
-                                  }
-                                }
-                                return (
-                                  <span key={r.id} className={`rounded-full px-2.5 py-1 text-xs font-medium ${darkMode ? 'bg-teal-700/80 text-teal-100' : 'bg-teal-600/80 text-white'}`}>
-                                    {timeLabel && <span className="opacity-70 mr-1">{timeLabel}</span>}{r.name}
-                                  </span>
-                                );
-                              })}
-                            </div>
-                          </div>
-                        );
-                      })()}
-
-                    </div>
+                    <GlanceSidebar variant="tablet" />
                   </div>
                 )}
 
@@ -1934,961 +972,8 @@ const DesktopLayout = () => {
               {/* Glance section — shown when glance tab active */}
               {tabletActiveTab === 'glance' && (
               <div className="p-4">
-                <div className="space-y-4">
-                  {/* Search bar + filter */}
-                  <div className="flex items-center gap-2">
-                    <button
-                      onClick={() => { setShowSpotlight(true); playUISound('spotlight'); }}
-                      className={`flex-1 flex items-center gap-2 px-3 py-2.5 rounded-lg ${darkMode ? 'bg-white/10 text-gray-400' : 'bg-black/5 text-stone-400'} transition-colors hover:opacity-80`}
-                    >
-                      <Search size={16} />
-                      <span className="text-sm">Search tasks...</span>
-                      <span className={`ml-auto text-xs ${textSecondary}`}>Ctrl+K</span>
-                    </button>
-                    {allTags.length > 0 && (
-                      <div className="flex-shrink-0 self-stretch flex items-center">
-                        <button
-                          ref={tagFilterBtnRef}
-                          onClick={() => setShowMobileTagFilter(v => !v)}
-                          className={`px-2.5 h-full flex items-center rounded-lg transition-colors ${
-                            !allTags.every(tag => selectedTags.includes(tag))
-                              ? 'bg-blue-500 text-white'
-                              : darkMode ? 'bg-white/10 text-gray-400' : 'bg-black/5 text-stone-400'
-                          } hover:opacity-80`}
-                        >
-                          <Filter size={16} />
-                        </button>
-                        {/* Desktop tag filter popover */}
-                        {showMobileTagFilter && (() => {
-                          const rect = tagFilterBtnRef.current?.getBoundingClientRect();
-                          return (
-                          <>
-                            <div className="fixed inset-0 z-40" onClick={() => setShowMobileTagFilter(false)} />
-                            <div
-                              className={`fixed z-50 ${cardBg} border ${borderClass} rounded-xl shadow-xl`}
-                              style={{ width: '280px', top: rect ? rect.bottom + 4 : 0, left: rect ? Math.max(8, rect.right - 280) : 0 }}
-                              onClick={(e) => e.stopPropagation()}
-                            >
-                              <div className="flex items-center justify-between px-3 py-2 border-b" style={{ borderColor: 'inherit' }}>
-                                <div className="flex items-center gap-1.5">
-                                  <Filter size={14} className={textSecondary} />
-                                  <span className={`text-sm font-semibold ${textPrimary}`}>Filter by Tag</span>
-                                </div>
-                                <div className="flex items-center gap-2">
-                                  {allTags.every(tag => selectedTags.includes(tag)) ? (
-                                    <button onClick={clearTagFilter} className="text-xs text-blue-500 hover:text-blue-600 font-medium">Clear</button>
-                                  ) : (
-                                    <button onClick={selectAllTags} className="text-xs text-blue-500 hover:text-blue-600 font-medium">Select All</button>
-                                  )}
-                                </div>
-                              </div>
-                              <div className="py-1 max-h-[300px] overflow-y-auto">
-                                {allTags.map(tag => {
-                                  const visibleDateStrs = new Set(visibleDates.map(d => dateToString(d)));
-                                  const regularCount = tasks.filter(t => !t.imported && visibleDateStrs.has(t.date) && extractTags(t.title).includes(tag)).length;
-                                  const recurringCount = expandedRecurringTasks.filter(t => visibleDateStrs.has(t.date) && extractTags(t.title).includes(tag)).length;
-                                  const tagCount = regularCount + recurringCount;
-                                  return (
-                                    <button
-                                      key={tag}
-                                      onClick={() => toggleTag(tag)}
-                                      className={`w-full flex items-center gap-2 px-3 py-2 transition-colors ${
-                                        tagCount === 0 ? 'opacity-40' : ''
-                                      } ${
-                                        selectedTags.includes(tag)
-                                          ? darkMode ? 'bg-blue-500/20' : 'bg-blue-50'
-                                          : darkMode ? 'hover:bg-white/5' : 'hover:bg-stone-50'
-                                      }`}
-                                    >
-                                      <div className={`w-4 h-4 rounded flex items-center justify-center flex-shrink-0 border transition-colors ${
-                                        selectedTags.includes(tag) ? 'bg-blue-500 border-blue-500' : darkMode ? 'border-gray-600' : 'border-stone-300'
-                                      }`}>
-                                        {selectedTags.includes(tag) && <Check size={12} className="text-white" />}
-                                      </div>
-                                      <Hash size={12} className={textSecondary} />
-                                      <span className={`flex-1 text-left text-sm ${textPrimary}`}>{tag}</span>
-                                      {tagCount > 0 && <span className={`text-xs ${textSecondary} tabular-nums`}>{tagCount}</span>}
-                                    </button>
-                                  );
-                                })}
-                              </div>
-                            </div>
-                          </>
-                          );
-                        })()}
-                      </div>
-                    )}
-                    {aiConfig.enabled && aiConfig.features.voiceTaskInput && (
-                      <button
-                        onClick={() => setShowVoiceInput(true)}
-                        className={`flex-shrink-0 self-stretch flex items-center px-2.5 rounded-lg transition-colors ${darkMode ? 'bg-white/10 text-purple-400' : 'bg-black/5 text-purple-600'} hover:opacity-80`}
-                        title="Voice Task Input (V)"
-                      >
-                        <Mic size={16} />
-                      </button>
-                    )}
-                  </div>
-
-                  {/* Morning dayGLANCE — AI morning summary card (desktop) */}
-                  {aiConfig.enabled && aiConfig.features.morningSummary && !morningGlanceDismissed && (
-                    (morningGlanceText || morningGlanceLoading || morningGlanceError) ? (
-                    <div className={`rounded-lg border p-3 ${darkMode ? 'border-amber-800/50 bg-amber-900/20' : 'border-amber-200 bg-amber-50'}`}>
-                      <div className="flex items-start justify-between gap-2">
-                        <div className="flex items-center gap-2 flex-shrink-0">
-                          <Sun size={16} className="text-amber-500" />
-                          <span className={`text-xs font-semibold uppercase tracking-wider ${darkMode ? 'text-amber-300' : 'text-amber-700'}`}>Morning Briefing</span>
-                        </div>
-                        <div className="flex items-center gap-1 flex-shrink-0">
-                          <button
-                            onClick={generateMorningSummary}
-                            className={`p-1 rounded transition-colors ${darkMode ? 'hover:bg-white/10' : 'hover:bg-black/5'}`}
-                            title="Regenerate"
-                          >
-                            <RefreshCw size={12} className={`${morningGlanceLoading ? 'animate-spin' : ''} ${textSecondary}`} />
-                          </button>
-                          <button
-                            onClick={dismissMorningGlance}
-                            className={`p-1 rounded transition-colors ${darkMode ? 'hover:bg-white/10' : 'hover:bg-black/5'}`}
-                            title="Dismiss for today"
-                          >
-                            <X size={12} className={textSecondary} />
-                          </button>
-                        </div>
-                      </div>
-                      <div className="mt-2">
-                        {morningGlanceLoading && (
-                          <div className="flex items-center gap-2">
-                            <Loader size={14} className={`animate-spin ${textSecondary}`} />
-                            <span className={`text-xs ${textSecondary}`}>Generating your briefing...</span>
-                          </div>
-                        )}
-                        {morningGlanceError && (
-                          <p className={`text-xs ${darkMode ? 'text-red-400' : 'text-red-600'}`}>{morningGlanceError}</p>
-                        )}
-                        {morningGlanceText && !morningGlanceLoading && (
-                          <p className={`text-sm leading-relaxed ${textPrimary}`}>{morningGlanceText}</p>
-                        )}
-                        {morningGlanceText && !morningGlanceLoading && aiConfig?.enabled && aiConfig.features?.smartScheduling && gtdFrames.filter(f => f.enabled).length > 0 && unscheduledTasks.filter(t => !t.completed && !t.isExample).length > 0 && (
-                          <button
-                            onClick={() => { setShowFramesModal(true); setFramesModalTab('schedule'); setEditingFrame(null); }}
-                            className={`mt-2 flex items-center gap-1.5 text-xs font-medium transition-colors ${darkMode ? 'text-purple-400 hover:text-purple-300' : 'text-purple-600 hover:text-purple-700'}`}
-                          >
-                            <BrainCircuit size={12} />
-                            Schedule inbox items?
-                          </button>
-                        )}
-                      </div>
-                    </div>
-                    ) : (
-                    <div
-                      className={`flex items-center gap-2 rounded-lg border px-3 py-2 cursor-pointer transition-colors ${darkMode ? 'border-amber-800/50 bg-amber-900/20 hover:bg-amber-900/30' : 'border-amber-200 bg-amber-50 hover:bg-amber-100'}`}
-                      onClick={generateMorningSummary}
-                    >
-                      <Sun size={14} className="text-amber-500 flex-shrink-0" />
-                      <span className={`text-sm ${darkMode ? 'text-amber-300' : 'text-amber-700'}`}>Click here to see AI briefing</span>
-                      <button
-                        onClick={(e) => { e.stopPropagation(); dismissMorningGlance(); }}
-                        className={`ml-auto p-0.5 rounded flex-shrink-0 transition-colors ${darkMode ? 'hover:bg-white/10' : 'hover:bg-black/10'}`}
-                        title="Dismiss for today"
-                      >
-                        <X size={12} className={textSecondary} />
-                      </button>
-                    </div>
-                    )
-                  )}
-
-                  {/* Evening Reflection — AI end-of-day card (desktop) */}
-                  {aiConfig.enabled && aiConfig.features.eveningReflection && !eveningGlanceDismissed && currentTime.getHours() >= 19 && (
-                    (eveningGlanceText || eveningGlanceLoading || eveningGlanceError) ? (
-                    <div className={`rounded-lg border p-3 ${darkMode ? 'border-indigo-800/50 bg-indigo-900/20' : 'border-indigo-200 bg-indigo-50'}`}>
-                      <div className="flex items-start justify-between gap-2">
-                        <div className="flex items-center gap-2 flex-shrink-0">
-                          <Moon size={16} className="text-indigo-400" />
-                          <span className={`text-xs font-semibold uppercase tracking-wider ${darkMode ? 'text-indigo-300' : 'text-indigo-700'}`}>Evening Reflection</span>
-                        </div>
-                        <div className="flex items-center gap-1 flex-shrink-0">
-                          <button onClick={generateEveningReflection} className={`p-1 rounded transition-colors ${darkMode ? 'hover:bg-white/10' : 'hover:bg-black/5'}`} title="Regenerate">
-                            <RefreshCw size={12} className={`${eveningGlanceLoading ? 'animate-spin' : ''} ${textSecondary}`} />
-                          </button>
-                          <button onClick={dismissEveningGlance} className={`p-1 rounded transition-colors ${darkMode ? 'hover:bg-white/10' : 'hover:bg-black/5'}`} title="Dismiss for today">
-                            <X size={12} className={textSecondary} />
-                          </button>
-                        </div>
-                      </div>
-                      <div className="mt-2">
-                        {eveningGlanceLoading && (
-                          <div className="flex items-center gap-2">
-                            <Loader size={14} className={`animate-spin ${textSecondary}`} />
-                            <span className={`text-xs ${textSecondary}`}>Reflecting on your day...</span>
-                          </div>
-                        )}
-                        {eveningGlanceError && <p className={`text-xs ${darkMode ? 'text-red-400' : 'text-red-600'}`}>{eveningGlanceError}</p>}
-                        {eveningGlanceText && !eveningGlanceLoading && <p className={`text-sm leading-relaxed ${textPrimary}`}>{eveningGlanceText}</p>}
-                        {eveningGlanceText && !eveningGlanceLoading && incompleteTodayTasks.length > 0 && gtdFrames.filter(f => f.enabled).length > 0 && aiConfig.features?.aiReschedule && (
-                          <button
-                            onClick={() => { setShowRescheduleModal(true); setRescheduleResults(null); setRescheduleError(''); }}
-                            className={`mt-2 flex items-center gap-1.5 text-xs font-medium transition-colors ${darkMode ? 'text-orange-400 hover:text-orange-300' : 'text-orange-600 hover:text-orange-700'}`}
-                          >
-                            <CalendarDays size={12} />
-                            Reschedule {incompleteTodayTasks.length} incomplete task{incompleteTodayTasks.length !== 1 ? 's' : ''} →
-                          </button>
-                        )}
-                      </div>
-                    </div>
-                    ) : (
-                    <div
-                      className={`flex items-center gap-2 rounded-lg border px-3 py-2 cursor-pointer transition-colors ${darkMode ? 'border-indigo-800/50 bg-indigo-900/20 hover:bg-indigo-900/30' : 'border-indigo-200 bg-indigo-50 hover:bg-indigo-100'}`}
-                      onClick={generateEveningReflection}
-                    >
-                      <Moon size={14} className="text-indigo-400 flex-shrink-0" />
-                      <span className={`text-sm ${darkMode ? 'text-indigo-300' : 'text-indigo-700'}`}>Click here for evening reflection</span>
-                      <button onClick={(e) => { e.stopPropagation(); dismissEveningGlance(); }} className={`ml-auto p-0.5 rounded flex-shrink-0 transition-colors ${darkMode ? 'hover:bg-white/10' : 'hover:bg-black/10'}`} title="Dismiss for today">
-                        <X size={12} className={textSecondary} />
-                      </button>
-                    </div>
-                    )
-                  )}
-
-                  {/* Frame Nudge card (desktop) */}
-                  {aiConfig.enabled && aiConfig.features?.frameNudge && activeFrameForNudge && activeFrameForNudge.minutesRemaining > 30 && agendaNowMarker.gapMinutes >= 15 && frameNudgeDismissedKey !== activeFrameNudgeKey && (frameNudgeSuggestion || frameNudgeLoading || frameNudgeError) && (
-                    <FrameNudgeCard
-                      suggestion={frameNudgeSuggestion}
-                      loading={frameNudgeLoading}
-                      error={frameNudgeError}
-                      activeFrame={activeFrameForNudge}
-                      darkMode={darkMode}
-                      textPrimary={textPrimary}
-                      textSecondary={textSecondary}
-                      onRefresh={generateFrameNudge}
-                      onDismiss={() => setFrameNudgeDismissedKey(activeFrameNudgeKey)}
-                      onStartTask={scheduleTaskAtNextSlot}
-                    />
-                  )}
-
-                  {/* Habit rings row */}
-                  {habitsEnabled && activeHabits.length > 0 && (
-                    <div className="relative">
-                      <div className="flex items-start gap-1 justify-center">
-                        {activeHabits.slice(0, 5).map((habit, habitIdx) => (
-                          <div key={habit.id} className="relative">
-                            <HabitRing
-                              size={44}
-                              habit={habit}
-                              count={getTodayHabitCount(habit.id)}
-                              darkMode={darkMode}
-                              onClick={() => incrementHabit(habit.id)}
-                              onContextMenu={(e) => { e.preventDefault(); setHabitLongPressId(prev => prev === habit.id ? null : habit.id); setHabitEditingCountId(null); }}
-                              onMouseDown={() => { if (habitLongPressTimer.current) clearTimeout(habitLongPressTimer.current); habitLongPressTimer.current = setTimeout(() => { setHabitLongPressId(prev => prev === habit.id ? null : habit.id); setHabitEditingCountId(null); }, 500); }}
-                              onMouseUp={() => { if (habitLongPressTimer.current) clearTimeout(habitLongPressTimer.current); }}
-                              onMouseLeave={() => { if (habitLongPressTimer.current) clearTimeout(habitLongPressTimer.current); }}
-                              onTouchStart={() => { if (habitLongPressTimer.current) clearTimeout(habitLongPressTimer.current); habitLongPressTimer.current = setTimeout(() => { setHabitLongPressId(prev => prev === habit.id ? null : habit.id); setHabitEditingCountId(null); }, 500); }}
-                              onTouchEnd={() => { if (habitLongPressTimer.current) clearTimeout(habitLongPressTimer.current); }}
-                            />
-                            {habitLongPressId === habit.id && (
-                              <>
-                                <div className="fixed inset-0 z-40" onClick={() => { setHabitLongPressId(null); setHabitEditingCountId(null); }} />
-                                <div className={`absolute top-full mt-1 z-50 ${darkMode ? 'bg-gray-800 border-gray-700' : 'bg-white border-stone-200'} border rounded-xl shadow-xl p-3 min-w-[140px] ${habitIdx === 0 ? 'left-0' : habitIdx === Math.min(activeHabits.length, 5) - 1 ? 'right-0' : 'left-1/2 -translate-x-1/2'}`}>
-                                  <div className={`text-xs font-semibold mb-2 text-center ${darkMode ? 'text-gray-300' : 'text-stone-700'}`}>{habit.name}</div>
-                                  <div className="flex items-center justify-center gap-3">
-                                    <button onClick={() => { setHabitCount(habit.id, getTodayHabitCount(habit.id) - 1); }} className={`w-8 h-8 rounded-full flex items-center justify-center ${darkMode ? 'bg-gray-700 text-gray-300 active:bg-gray-600' : 'bg-stone-100 text-stone-600 active:bg-stone-200'}`}><Minus size={16} /></button>
-                                    {habitEditingCountId === habit.id ? (
-                                    <input
-                                      type="number"
-                                      autoFocus
-                                      defaultValue={getTodayHabitCount(habit.id)}
-                                      onBlur={(e) => { setHabitCount(habit.id, parseInt(e.target.value) || 0); setHabitEditingCountId(null); }}
-                                      onKeyDown={(e) => { if (e.key === 'Enter') e.target.blur(); }}
-                                      onClick={(e) => e.stopPropagation()}
-                                      className={`w-16 text-lg font-bold text-center rounded-lg border ${darkMode ? 'bg-gray-700 text-white border-gray-600' : 'bg-stone-50 text-stone-900 border-stone-300'} outline-none focus:ring-2 focus:ring-blue-500 [appearance:textfield] [&::-webkit-inner-spin-button]:appearance-none [&::-webkit-outer-spin-button]:appearance-none`}
-                                      onFocus={(e) => e.target.select()}
-                                    />
-                                  ) : (
-                                    <span onClick={(e) => { e.stopPropagation(); setHabitEditingCountId(habit.id); }} className={`text-lg font-bold min-w-[2ch] text-center cursor-pointer hover:opacity-70 ${darkMode ? 'text-white' : 'text-stone-900'}`}>{getTodayHabitCount(habit.id)}</span>
-                                  )}
-                                    <button onClick={() => { incrementHabit(habit.id); }} className={`w-8 h-8 rounded-full flex items-center justify-center ${darkMode ? 'bg-gray-700 text-gray-300 active:bg-gray-600' : 'bg-stone-100 text-stone-600 active:bg-stone-200'}`}><Plus size={16} /></button>
-                                  </div>
-                                  <button onClick={() => { setHabitCount(habit.id, 0); setHabitLongPressId(null); setHabitEditingCountId(null); }} className="mt-2 w-full text-xs text-red-500 font-medium py-1 rounded hover:bg-red-500/10 transition-colors">Reset</button>
-                                </div>
-                              </>
-                            )}
-                          </div>
-                        ))}
-                        {activeHabits.length > 5 && (
-                          <div className="relative">
-                            <button
-                              onClick={() => setHabitOverflowOpen(prev => !prev)}
-                              className={`w-[52px] h-[44px] flex items-center justify-center rounded-lg text-xs font-bold ${darkMode ? 'bg-gray-700 text-gray-400 active:bg-gray-600' : 'bg-stone-100 text-stone-500 active:bg-stone-200'} transition-colors`}
-                            >
-                              +{activeHabits.length - 5}
-                            </button>
-                            {habitOverflowOpen && (
-                              <>
-                                <div className="fixed inset-0 z-40" onClick={() => setHabitOverflowOpen(false)} />
-                                <div className={`absolute top-full right-0 mt-1 z-50 ${darkMode ? 'bg-gray-800 border-gray-700' : 'bg-white border-stone-200'} border rounded-xl shadow-xl p-2 min-w-[180px]`}>
-                                  {activeHabits.slice(5).map(habit => {
-                                    const count = getTodayHabitCount(habit.id);
-                                    const IconComp = HABIT_ICONS[habit.icon] || Target;
-                                    const colorObj = HABIT_COLORS.find(c => c.name === habit.color) || HABIT_COLORS[0];
-                                    return (
-                                      <button
-                                        key={habit.id}
-                                        onClick={() => { incrementHabit(habit.id); }}
-                                        className={`w-full flex items-center gap-2 px-2 py-1.5 rounded-lg ${darkMode ? 'hover:bg-gray-700 active:bg-gray-600' : 'hover:bg-stone-50 active:bg-stone-100'} transition-colors`}
-                                      >
-                                        <IconComp size={16} style={{ color: colorObj.ring }} />
-                                        <span className={`text-sm flex-1 text-left ${darkMode ? 'text-gray-300' : 'text-stone-700'}`}>{habit.name}</span>
-                                        <span className={`text-xs font-semibold ${darkMode ? 'text-gray-400' : 'text-stone-500'}`}>{count}/{habit.target}</span>
-                                      </button>
-                                    );
-                                  })}
-                                </div>
-                              </>
-                            )}
-                          </div>
-                        )}
-                      </div>
-                    </div>
-                  )}
-                  {/* Getting Started checklist */}
-                  {showGettingStarted && <GettingStartedChecklist
-                  items={gettingStartedItems}
-                  completedCount={gettingStartedCompleteCount}
-                  darkMode={darkMode}
-                  textPrimary={textPrimary}
-                  textSecondary={textSecondary}
-                  onDismiss={() => setGettingStartedDismissed(true)}
-                  onComplete={() => {
-                    setTasks(prev => prev.filter(t => !t.isExample));
-                    setUnscheduledTasks(prev => prev.filter(t => !t.isExample));
-                    setRecycleBin(prev => prev.filter(t => !t.isExample));
-                    setRecurringTasks(prev => prev.filter(t => !t.isExample));
-                    setOnboardingComplete(true);
-                    setGettingStartedDismissed(true);
-                  }}
-                />}
-                  {/* Reschedule Tasks — shown when overdue past-day tasks exist, or incomplete today tasks after 7pm */}
-                  {aiConfig?.enabled && aiConfig.features?.aiReschedule && gtdFrames.filter(f => f.enabled).length > 0 && (() => {
-                    const _todayStr = getTodayStr();
-                    const _pastOverdue = getOverdueTasks().filter(t => t._overdueType === 'scheduled' ? t.date < _todayStr : true);
-                    if (!(_pastOverdue.length > 0 || (incompleteTodayTasks.length > 0 && currentTime.getHours() >= 19))) return null;
-                    return (
-                      <button
-                        onClick={() => { setShowRescheduleModal(true); setRescheduleResults(null); setRescheduleError(''); }}
-                        className={`w-full mb-3 flex items-center justify-center gap-2 py-2.5 px-4 rounded-xl text-sm font-semibold transition-colors ${darkMode ? 'bg-violet-600/20 hover:bg-violet-600/30 text-violet-300 border border-violet-500/30' : 'bg-violet-50 hover:bg-violet-100 text-violet-700 border border-violet-200'}`}
-                      >
-                        <Sparkles size={15} />
-                        Reschedule Tasks
-                      </button>
-                    );
-                  })()}
-                  {/* Overdue tasks from past days — matching tablet landscape */}
-                  {(() => {
-                    const todayStr = getTodayStr();
-                    const pastOverdue = getOverdueTasks().filter(t => {
-                      if (t._overdueType === 'scheduled') return t.date < todayStr;
-                      return true;
-                    });
-                    if (pastOverdue.length === 0) return null;
-                    return (
-                      <div className={`rounded-lg border ${darkMode ? 'border-orange-500/40 bg-orange-500/10' : 'border-orange-400/50 bg-orange-50'} overflow-hidden`}>
-                        <button
-                          onClick={() => toggleSection('overdue')}
-                          className="w-full flex items-center justify-between px-3 py-2.5"
-                        >
-                          <div className="flex items-center gap-2">
-                            <AlertTriangle size={15} className="text-orange-500" />
-                            <span className="text-sm font-semibold text-orange-500">Overdue</span>
-                            <span className={`text-xs px-1.5 py-0.5 rounded-full ${darkMode ? 'bg-orange-500/30 text-orange-300' : 'bg-orange-200 text-orange-700'}`}>
-                              {pastOverdue.length}
-                            </span>
-                          </div>
-                          {minimizedSections.overdue ? <ChevronDown size={16} className="text-orange-500" /> : <ChevronUp size={16} className="text-orange-500" />}
-                        </button>
-                        {!minimizedSections.overdue && (
-                          <div className="px-3 pb-2.5 space-y-1">
-                            {pastOverdue.map(task => (
-                              <div
-                                key={`desktop-overdue-${task.id}`}
-                                className={`flex items-center gap-2.5 py-2 px-2 rounded-lg ${darkMode ? 'bg-white/5' : 'bg-white/80'}`}
-                              >
-                                <button
-                                  onClick={() => toggleComplete(task.id, task._overdueType === 'deadline')}
-                                  className={`w-5 h-5 rounded flex-shrink-0 border-2 ${task.completed
-                                    ? 'border-orange-400 bg-orange-400'
-                                    : darkMode ? 'border-orange-400/60 bg-white/10' : 'border-orange-400/60 bg-white'
-                                  } flex items-center justify-center`}
-                                >
-                                  {task.completed && <Check size={12} strokeWidth={3} className="text-white" />}
-                                </button>
-                                <div
-                                  className={`flex-1 min-w-0 ${task._overdueType === 'scheduled' ? 'cursor-pointer' : ''}`}
-                                  onClick={() => {
-                                    if (task._overdueType !== 'scheduled') return;
-                                    if (task.date) setSelectedDate(new Date(task.date + 'T12:00:00'));
-                                    setTimeout(() => {
-                                      const el = document.querySelector(`[data-task-id="${task.id}"]`);
-                                      if (el && calendarRef.current) {
-                                        const container = calendarRef.current;
-                                        const elTop = el.getBoundingClientRect().top - container.getBoundingClientRect().top + container.scrollTop;
-                                        const scrollTarget = Math.max(0, elTop - container.clientHeight / 2 + el.offsetHeight / 2);
-                                        container.scrollTo({ top: scrollTarget, behavior: 'smooth' });
-                                        el.classList.add('ring-2', 'ring-blue-400');
-                                        setTimeout(() => el.classList.remove('ring-2', 'ring-blue-400'), 2000);
-                                      }
-                                    }, 200);
-                                  }}
-                                >
-                                  <div className={`text-sm font-medium truncate ${task.completed ? 'line-through opacity-50' : textPrimary}`}>
-                                    {renderTitle(task.title)}
-                                  </div>
-                                  <div className={`text-xs ${textSecondary} flex items-center gap-1 mt-0.5`}>
-                                    {task._overdueType === 'scheduled' ? (
-                                      <><CalendarDays size={10} /> {formatDeadlineDate(task.date)}</>
-                                    ) : (
-                                      <><AlertCircle size={10} /> Due: {formatDeadlineDate(task.deadline)}</>
-                                    )}
-                                  </div>
-                                </div>
-                                <div className="flex items-center gap-1 flex-shrink-0">
-                                  {task.isRecurring ? (
-                                    <button
-                                      onClick={() => toggleComplete(task.id, false)}
-                                      className={`p-1.5 rounded-lg ${darkMode ? 'bg-white/10 text-gray-400' : 'bg-stone-100 text-stone-500'} hover:scale-95 transition-transform`}
-                                      title="Mark complete"
-                                    >
-                                      <CheckCircle size={14} />
-                                    </button>
-                                  ) : (
-                                  <button
-                                    onClick={() => {
-                                      if (task._overdueType === 'scheduled') {
-                                        pushUndo();
-                                        setTasks(prev => prev.filter(t => t.id !== task.id));
-                                        const { startTime, date, duration, _overdueType, ...rest } = task;
-                                        setUnscheduledTasks(prev => [...prev, { ...rest, priority: rest.priority || 0 }]);
-                                        playUISound('slide');
-                                        setUndoToast({ message: 'Moved to inbox', actionable: true });
-                                      } else {
-                                        clearDeadline(task.id);
-                                        playUISound('slide');
-                                        setUndoToast({ message: 'Deadline cleared', actionable: true });
-                                      }
-                                    }}
-                                    className={`p-1.5 rounded-lg ${darkMode ? 'bg-white/10 text-gray-400' : 'bg-stone-100 text-stone-500'} hover:scale-95 transition-transform`}
-                                    title="Move to inbox"
-                                  >
-                                    <Inbox size={14} />
-                                  </button>
-                                  )}
-                                  <button
-                                    onClick={() => moveToRecycleBin(task.id, task._overdueType === 'deadline')}
-                                    className={`p-1.5 rounded-lg ${darkMode ? 'bg-white/10 text-gray-400' : 'bg-stone-100 text-stone-500'} hover:scale-95 transition-transform`}
-                                    title="Move to Recycle Bin"
-                                  >
-                                    <Trash2 size={14} />
-                                  </button>
-                                </div>
-                              </div>
-                            ))}
-                          </div>
-                        )}
-                      </div>
-                    );
-                  })()}
-
-                  {/* Today's agenda — grouped by frames, matching tablet landscape */}
-                  {(() => {
-                    const filteredAgenda = filterByTags(todayAgenda);
-                    const today = new Date(getTodayStr() + 'T12:00:00');
-                    const nowMinGlance = currentTime.getHours() * 60 + currentTime.getMinutes();
-                    const todayFrames = getFrameInstancesForDate(today).filter(f => timeToMinutes(f.end) > nowMinGlance);
-                    const glanceBorderColorMap = darkMode ? {
-                      'bg-indigo-200': 'rgba(165,180,252,0.4)',
-                      'bg-amber-200': 'rgba(253,230,138,0.4)',
-                      'bg-green-200': 'rgba(167,243,208,0.4)',
-                      'bg-blue-200': 'rgba(191,219,254,0.4)',
-                      'bg-rose-200': 'rgba(254,205,211,0.4)',
-                      'bg-purple-200': 'rgba(221,214,254,0.4)',
-                      'bg-teal-200': 'rgba(153,246,228,0.4)',
-                      'bg-orange-200': 'rgba(254,215,170,0.4)',
-                    } : {
-                      'bg-indigo-200': 'rgba(79,70,229,0.75)',
-                      'bg-amber-200': 'rgba(217,119,6,0.75)',
-                      'bg-green-200': 'rgba(22,163,74,0.75)',
-                      'bg-blue-200': 'rgba(37,99,235,0.75)',
-                      'bg-rose-200': 'rgba(225,29,72,0.75)',
-                      'bg-purple-200': 'rgba(147,51,234,0.75)',
-                      'bg-teal-200': 'rgba(13,148,136,0.75)',
-                      'bg-orange-200': 'rgba(234,88,12,0.75)',
-                    };
-                    const glanceColorMap = darkMode ? {
-                      'bg-indigo-200': 'rgba(165,180,252,0.08)',
-                      'bg-amber-200': 'rgba(253,230,138,0.08)',
-                      'bg-green-200': 'rgba(167,243,208,0.08)',
-                      'bg-blue-200': 'rgba(191,219,254,0.08)',
-                      'bg-rose-200': 'rgba(254,205,211,0.08)',
-                      'bg-purple-200': 'rgba(221,214,254,0.08)',
-                      'bg-teal-200': 'rgba(153,246,228,0.08)',
-                      'bg-orange-200': 'rgba(254,215,170,0.08)',
-                    } : {
-                      'bg-indigo-200': 'rgba(165,180,252,0.18)',
-                      'bg-amber-200': 'rgba(253,230,138,0.18)',
-                      'bg-green-200': 'rgba(167,243,208,0.18)',
-                      'bg-blue-200': 'rgba(191,219,254,0.18)',
-                      'bg-rose-200': 'rgba(254,205,211,0.18)',
-                      'bg-purple-200': 'rgba(221,214,254,0.18)',
-                      'bg-teal-200': 'rgba(153,246,228,0.18)',
-                      'bg-orange-200': 'rgba(254,215,170,0.18)',
-                    };
-
-                    // Classify each agenda task into a frame or "unframed"
-                    const taskFrameMap = new Map();
-                    for (const task of filteredAgenda) {
-                      if (task._agendaType !== 'scheduled' || !task.startTime) continue;
-                      const tStart = timeToMinutes(task.startTime);
-                      const tEnd = tStart + (task.duration || 30);
-                      for (const frame of todayFrames) {
-                        const fStart = timeToMinutes(frame.start);
-                        const fEnd = timeToMinutes(frame.end);
-                        if (tStart >= fStart && tEnd <= fEnd) {
-                          taskFrameMap.set(String(task.id), frame.frameId);
-                          break;
-                        }
-                      }
-                    }
-
-                    const nonScheduled = filteredAgenda.filter(t => t._agendaType !== 'scheduled');
-                    const scheduled = filteredAgenda.filter(t => t._agendaType === 'scheduled');
-
-                    const sections = [];
-                    const framedIds = new Set(taskFrameMap.values());
-                    const sortedFrames = [...todayFrames].filter(f => framedIds.has(f.frameId) || true).sort((a, b) => timeToMinutes(a.start) - timeToMinutes(b.start));
-                    const assignedTaskIds = new Set();
-
-                    let scheduledIdx = 0;
-                    for (const frame of sortedFrames) {
-                      const fStart = timeToMinutes(frame.start);
-                      const beforeTasks = [];
-                      while (scheduledIdx < scheduled.length) {
-                        const t = scheduled[scheduledIdx];
-                        const tStart = timeToMinutes(t.startTime || '00:00');
-                        if (tStart < fStart && !taskFrameMap.has(String(t.id))) {
-                          beforeTasks.push(t);
-                          assignedTaskIds.add(String(t.id));
-                          scheduledIdx++;
-                        } else break;
-                      }
-                      if (beforeTasks.length > 0) sections.push({ type: 'unframed', tasks: beforeTasks });
-
-                      const frameTasks = scheduled.filter(t => taskFrameMap.get(String(t.id)) === frame.frameId);
-                      const availSlots = computeAvailableSlots(frame, today);
-                      const totalAvail = availSlots.reduce((sum, s) => sum + s.minutes, 0);
-                      // Hide frames with no availability and no tasks (fully blocked by calendar events)
-                      if (totalAvail > 0 || frameTasks.length > 0) {
-                        sections.push({ type: 'frame', frame, tasks: frameTasks, totalAvail });
-                      }
-                      frameTasks.forEach(t => assignedTaskIds.add(String(t.id)));
-                      while (scheduledIdx < scheduled.length && assignedTaskIds.has(String(scheduled[scheduledIdx].id))) scheduledIdx++;
-                    }
-                    const remaining = scheduled.filter(t => !assignedTaskIds.has(String(t.id)));
-                    if (remaining.length > 0) sections.push({ type: 'unframed', tasks: remaining });
-
-                    const nowMin = currentTime.getHours() * 60 + currentTime.getMinutes();
-
-                    // Compute now marker position within sections (frame-aware)
-                    let nowMarkerSectionInfo = null;
-                    let nowIsAfterAllSections = false;
-                    if (sections.length > 0 && !agendaNowMarker.insideTask) {
-                      for (let si = 0; si < sections.length; si++) {
-                        const section = sections[si];
-                        let sStart, sEnd;
-                        if (section.type === 'frame') {
-                          sStart = timeToMinutes(section.frame.start);
-                          sEnd = timeToMinutes(section.frame.end);
-                        } else {
-                          if (section.tasks.length === 0) continue;
-                          sStart = Math.min(...section.tasks.map(t => timeToMinutes(t.startTime || '00:00')));
-                          sEnd = Math.max(...section.tasks.map(t => timeToMinutes(t.startTime || '00:00') + (t.duration || 0)));
-                        }
-                        if (nowMin >= sStart && nowMin < sEnd) {
-                          let afterTaskIdx = -1;
-                          for (let ti = 0; ti < section.tasks.length; ti++) {
-                            const taskEnd = timeToMinutes(section.tasks[ti].startTime || '00:00') + (section.tasks[ti].duration || 0);
-                            if (nowMin >= taskEnd) afterTaskIdx = ti;
-                          }
-                          nowMarkerSectionInfo = { si, inSection: true, afterTaskIdx };
-                          break;
-                        } else if (nowMin < sStart) {
-                          nowMarkerSectionInfo = { si, inSection: false };
-                          break;
-                        }
-                      }
-                      if (!nowMarkerSectionInfo) {
-                        nowIsAfterAllSections = true;
-                      }
-                    }
-
-                    const renderDesktopNowMarker = (key) => {
-                      const gapH = Math.floor(agendaNowMarker.gapMinutes / 60);
-                      const gapM = agendaNowMarker.gapMinutes % 60;
-                      const gapStr = gapH > 0 ? `${gapH}h${gapM > 0 ? ` ${gapM}m` : ''}` : `${gapM}m`;
-                      return (
-                        <div key={key} className="flex gap-2.5 py-2">
-                          <div className="w-1.5 rounded-full flex-shrink-0 bg-red-500" />
-                          <div className="min-w-0 flex-1">
-                            <div className="text-sm font-medium text-red-500">{formatTime(agendaNowMarker.nowTimeStr)}, {gapStr} of free time</div>
-                            {agendaNowMarker.gapMinutes < 30 ? (
-                              <div className="text-xs italic text-red-500 mt-0.5">Get ready to be productive!</div>
-                            ) : agendaNowMarker.inboxCount > 0 ? (
-                              <div className="flex items-center gap-1.5 mt-0.5">
-                                <span className="text-xs italic text-red-500">Maybe tackle an inbox task?</span>
-                                {agendaNowMarker.showNudge && aiConfig.enabled && aiConfig.features?.frameNudge && activeFrameForNudge && (
-                                  <button onClick={() => { setFrameNudgeDismissedKey(''); generateFrameNudge(); }} className="flex items-center gap-0.5 px-1.5 py-0.5 rounded text-[10px] font-medium bg-teal-500/20 hover:bg-teal-500/30 text-teal-600 dark:text-teal-400 transition-colors">
-                                    <Sparkles size={9} />AI
-                                  </button>
-                                )}
-                              </div>
-                            ) : null}
-                          </div>
-                        </div>
-                      );
-                    };
-
-                    const renderTaskItem = (task, keyPrefix) => {
-                      const colorClass = task.color === 'task-calendar' ? '' : task.color;
-                      let timeLabel = '';
-                      let relativeLabel = '';
-                      if (task._agendaType === 'allday') {
-                        timeLabel = 'ALL DAY';
-                      } else if (task._agendaType === 'deadline') {
-                        timeLabel = 'DUE TODAY';
-                      } else {
-                        const [h, m] = (task.startTime || '0:0').split(':').map(Number);
-                        const startMin = h * 60 + m;
-                        const endMin = startMin + (task.duration || 0);
-                        const endH = String(Math.floor(endMin / 60)).padStart(2, '0');
-                        const endM = String(endMin % 60).padStart(2, '0');
-                        timeLabel = `${formatTime(task.startTime)}\u00A0–\u00A0${formatTime(endH + ':' + endM)}`;
-                        const diff = startMin - nowMin;
-                        if (diff > 0) {
-                          relativeLabel = diff >= 60 ? `in ${Math.floor(diff / 60)}h ${diff % 60 > 0 ? `${diff % 60}m` : ''}` : `in ${diff}m`;
-                        } else if (diff === 0) {
-                          relativeLabel = 'now';
-                        } else if (nowMin < endMin && !task.completed) {
-                          relativeLabel = 'In Progress';
-                        } else if (nowMin >= endMin && !task.completed) {
-                          relativeLabel = 'Overdue';
-                        }
-                      }
-                      return (
-                        <div
-                          key={`${keyPrefix}-${task._agendaType}-${task.id}`}
-                          data-ctx-menu
-                          onContextMenu={(e) => {
-                            e.preventDefault();
-                            setTaskContextMenu({
-                              x: e.clientX, y: e.clientY,
-                              taskId: task.id,
-                              isRecurring: !!task.isRecurring,
-                              isImported: !!task.imported,
-                              isAllDay: !!task.isAllDay || task._agendaType === 'allday',
-                              dateStr: dateToString(new Date()),
-                            });
-                          }}
-                          className={`flex gap-2.5 py-2 ${task.completed ? 'opacity-50' : ''} cursor-pointer hover:bg-white/5 rounded-lg transition-colors`}
-                          onClick={() => {
-                            const el = document.querySelector(`[data-task-id="${task.id}"]`);
-                            if (el && calendarRef.current) {
-                              const container = calendarRef.current;
-                              const elTop = el.getBoundingClientRect().top - container.getBoundingClientRect().top + container.scrollTop;
-                              const scrollTarget = Math.max(0, elTop - container.clientHeight / 2 + el.offsetHeight / 2);
-                              container.scrollTo({ top: scrollTarget, behavior: 'smooth' });
-                              el.classList.add('ring-2', 'ring-blue-400');
-                              setTimeout(() => el.classList.remove('ring-2', 'ring-blue-400'), 2000);
-                            }
-                          }}
-                        >
-                          <div className={`w-1.5 rounded-full flex-shrink-0 ${colorClass} ${relativeLabel === 'In Progress' ? 'animate-pulse' : ''}`} style={task.isTaskCalendar ? getTaskCalendarStyle(task, darkMode) : task.nativeCalendarColor ? { backgroundColor: task.nativeCalendarColor } : {}}></div>
-                          <div className="min-w-0 flex-1">
-                            <div className={`text-sm font-semibold ${textPrimary} ${task.completed ? 'line-through' : ''} flex items-center gap-1.5`}>
-                              {task.isRecurring && <RefreshCw size={13} className="flex-shrink-0 opacity-60" />}
-                              {task.importSource === 'obsidian' && <BookOpen size={13} className="flex-shrink-0 opacity-60" title="From Obsidian" />}
-                              <span className="truncate">{renderTitle(task.title)}</span>
-                              {isNativeAndroid() && extractWikilinks(task.title).map((note, i) => (
-                                <button key={i} className="flex-shrink-0 text-purple-400 active:text-purple-300" onClick={(e) => { e.stopPropagation(); window.DayGlanceObsidian?.openNote(note); }} title={`Open "${note}" in Obsidian`}><NotebookPen size={13} /></button>
-                              ))}
-                            </div>
-                            <div className={`text-sm ${textSecondary} flex items-center gap-1`}>
-                              <span className="whitespace-nowrap">{timeLabel}{relativeLabel ? ',' : ''}</span>{relativeLabel ? <span className={relativeLabel === 'Overdue' ? 'text-orange-500 font-medium' : relativeLabel === 'In Progress' ? 'text-blue-500 font-medium' : ''}>{relativeLabel}</span> : ''}
-                              {relativeLabel === 'In Progress' && focusModeAvailable && (
-                                <button
-                                  onClick={(e) => { e.stopPropagation(); enterFocusMode(); }}
-                                  className="ml-1 p-1.5 rounded text-purple-500 hover:text-purple-400 hover:bg-purple-500/20 transition-colors"
-                                  title="Enter Focus Mode"
-                                >
-                                  <Target size={16} className="animate-pulse" />
-                                </button>
-                              )}
-                            </div>
-                            {goalsProjectsEnabled && task.projectId && (() => {
-                              const proj = projects.find(p => p.id === task.projectId);
-                              if (!proj) return null;
-                              return (
-                                <button
-                                  onClick={(e) => { e.stopPropagation(); setProjectFilter(prev => prev === task.projectId ? null : task.projectId); }}
-                                  className={`inline-flex items-center text-[10px] px-1.5 py-0.5 rounded-full font-medium transition-colors ${darkMode ? 'bg-blue-900/50 text-blue-300 hover:bg-blue-800/70' : 'bg-blue-100 text-blue-700 hover:bg-blue-200'} ${projectFilter === task.projectId ? 'ring-1 ring-blue-400' : ''}`}
-                                  title={projectFilter === task.projectId ? 'Clear project filter' : `Filter: ${proj.title}`}
-                                >
-                                  {proj.title}
-                                </button>
-                              );
-                            })()}
-                          </div>
-                          {(relativeLabel === 'Overdue' || (task._agendaType === 'allday' && !task.imported)) && !task.completed && (
-                            <div className="flex items-center gap-1 flex-shrink-0 mr-5">
-                              {!task.isRecurring && (
-                                <button
-                                  onClick={(e) => {
-                                    e.stopPropagation();
-                                    pushUndo();
-                                    setTasks(prev => prev.filter(t => t.id !== task.id));
-                                    const { startTime, date, _agendaType, ...rest } = task;
-                                    setUnscheduledTasks(prev => [...prev, { ...rest, priority: rest.priority || 0 }]);
-                                    playUISound('slide');
-                                    setUndoToast({ message: 'Moved to inbox', actionable: true });
-                                  }}
-                                  className={`p-1.5 rounded-lg ${darkMode ? 'bg-white/10 text-gray-400' : 'bg-stone-100 text-stone-500'} hover:scale-95 transition-transform`}
-                                  title="Move to Inbox"
-                                >
-                                  <Inbox size={14} />
-                                </button>
-                              )}
-                              <button
-                                onClick={(e) => { e.stopPropagation(); toggleComplete(task.id, false); }}
-                                className={`p-1.5 rounded-lg ${darkMode ? 'bg-white/10 text-gray-400' : 'bg-stone-100 text-stone-500'} hover:scale-95 transition-transform`}
-                                title="Mark complete"
-                              >
-                                <CheckCircle size={14} />
-                              </button>
-                            </div>
-                          )}
-                        </div>
-                      );
-                    };
-
-                    return (
-                    <div className="space-y-1.5">
-                      {filteredAgenda.length === 0 && (
-                        <p className={`text-sm ${textSecondary} text-center py-4`}>No tasks scheduled for today</p>
-                      )}
-                      {/* Now marker before first task (only when no frame sections handle positioning) */}
-                      {filteredAgenda.length > 0 && sections.length === 0 && !agendaNowMarker.insideTask && agendaNowMarker.insertAfterIndex < 0 && (() => {
-                        const gapH = Math.floor(agendaNowMarker.gapMinutes / 60);
-                        const gapM = agendaNowMarker.gapMinutes % 60;
-                        const gapStr = gapH > 0 ? `${gapH}h${gapM > 0 ? ` ${gapM}m` : ''}` : `${gapM}m`;
-                        return (
-                          <div key="desktop-now-marker" className="flex gap-2.5 py-2.5">
-                            <div className="w-1.5 rounded-full flex-shrink-0 bg-red-500" />
-                            <div className="min-w-0 flex-1">
-                              <div className="text-sm font-medium text-red-500">{formatTime(agendaNowMarker.nowTimeStr)}, {gapStr} of free time</div>
-                              {agendaNowMarker.gapMinutes < 30 ? (
-                                <div className="text-xs italic text-red-500 mt-0.5">Get ready to be productive!</div>
-                              ) : agendaNowMarker.inboxCount > 0 ? (
-                                <div className="flex items-center gap-1.5 mt-0.5">
-                                  <span className="text-xs italic text-red-500">Maybe tackle an inbox task?</span>
-                                  {agendaNowMarker.showNudge && aiConfig.enabled && aiConfig.features?.frameNudge && activeFrameForNudge && (
-                                    <button onClick={() => { setFrameNudgeDismissedKey(''); generateFrameNudge(); }} className="flex items-center gap-0.5 px-1.5 py-0.5 rounded text-[10px] font-medium bg-teal-500/20 hover:bg-teal-500/30 text-teal-600 dark:text-teal-400 transition-colors">
-                                      <Sparkles size={9} />AI
-                                    </button>
-                                  )}
-                                </div>
-                              ) : null}
-                            </div>
-                          </div>
-                        );
-                      })()}
-                      {/* Non-scheduled items (all-day, deadlines) */}
-                      {nonScheduled.map(task => renderTaskItem(task, 'desktop-glance'))}
-                      {/* Frame-grouped and unframed sections */}
-                      {sections.map((section, si) => {
-                        const elements = [];
-                        // Now marker between sections (before this section)
-                        if (nowMarkerSectionInfo && !nowMarkerSectionInfo.inSection && nowMarkerSectionInfo.si === si) {
-                          elements.push(renderDesktopNowMarker(`desktop-now-mid-${si}`));
-                        }
-                        if (section.type === 'frame') {
-                          const borderColor = glanceBorderColorMap[section.frame.color] || (darkMode ? 'rgba(165,180,252,0.4)' : 'rgba(79,70,229,0.75)');
-                          const bgColor = glanceColorMap[section.frame.color] || (darkMode ? 'rgba(165,180,252,0.08)' : 'rgba(165,180,252,0.18)');
-                          const availH = Math.floor(section.totalAvail / 60);
-                          const availM = section.totalAvail % 60;
-                          const availStr = availH > 0 ? `${availH}h${availM > 0 ? ` ${availM}m` : ''}` : `${availM}m`;
-                          const markerInThisFrame = nowMarkerSectionInfo && nowMarkerSectionInfo.inSection && nowMarkerSectionInfo.si === si;
-                          elements.push(
-                            <div
-                              key={`desktop-frame-section-${section.frame.frameId}`}
-                              className="rounded-md overflow-hidden"
-                              style={{
-                                borderLeft: `3px solid ${borderColor}`,
-                                background: bgColor,
-                              }}
-                            >
-                              <div className="px-3 pt-2 pb-1">
-                                <div className="flex items-center gap-1.5">
-                                  <LayoutGrid size={12} style={{ color: borderColor }} />
-                                  <span className="text-xs font-semibold" style={{ color: borderColor }}>{section.frame.label}</span>
-                                  <span className={`text-xs ${textSecondary}`}>{formatTime(section.frame.start)} – {formatTime(section.frame.end)}</span>
-                                </div>
-                                {section.totalAvail > 0 && (
-                                  <p className="mt-1">
-                                    <span className={`text-xs font-medium px-2 py-0.5 rounded-full ${darkMode ? 'bg-blue-900/40 text-blue-300' : 'bg-blue-100 text-blue-700'}`}>
-                                      {availStr} available
-                                    </span>
-                                  </p>
-                                )}
-                              </div>
-                              <div className="px-2 pb-1.5">
-                                {section.tasks.length === 0 && !markerInThisFrame ? (
-                                  <p className={`text-xs ${textSecondary} py-2 px-1 italic`}>No tasks scheduled</p>
-                                ) : (() => {
-                                  const items = [];
-                                  if (markerInThisFrame && nowMarkerSectionInfo.afterTaskIdx < 0) {
-                                    items.push(renderDesktopNowMarker(`desktop-now-frame-${si}`));
-                                  }
-                                  section.tasks.forEach((task, ti) => {
-                                    items.push(renderTaskItem(task, 'desktop-frame'));
-                                    if (markerInThisFrame && nowMarkerSectionInfo.afterTaskIdx === ti) {
-                                      items.push(renderDesktopNowMarker(`desktop-now-frame-${si}-${ti}`));
-                                    }
-                                  });
-                                  return items;
-                                })()}
-                              </div>
-                            </div>
-                          );
-                        } else {
-                          const markerInThisSection = nowMarkerSectionInfo && nowMarkerSectionInfo.inSection && nowMarkerSectionInfo.si === si;
-                          if (markerInThisSection && nowMarkerSectionInfo.afterTaskIdx < 0) {
-                            elements.push(renderDesktopNowMarker(`desktop-now-unframed-${si}`));
-                          }
-                          section.tasks.forEach((task, ti) => {
-                            elements.push(renderTaskItem(task, 'desktop-glance'));
-                            if (markerInThisSection && nowMarkerSectionInfo.afterTaskIdx === ti) {
-                              elements.push(renderDesktopNowMarker(`desktop-now-unframed-${si}-${ti}`));
-                            }
-                          });
-                        }
-                        return <React.Fragment key={`desktop-section-${si}`}>{elements}</React.Fragment>;
-                      })}
-                      {/* Now marker after all tasks/frames */}
-                      {filteredAgenda.length > 0 && !agendaNowMarker.insideTask && (sections.length > 0 ? nowIsAfterAllSections : agendaNowMarker.insertAfterIndex >= todayAgenda.length - 1) && (() => {
-                        const hr = currentTime.getHours();
-                        const barColor = hr >= 22 ? 'bg-blue-500' : hr >= 19 ? 'bg-green-500' : 'bg-yellow-500';
-                        const textColor = hr >= 22 ? 'text-blue-500' : hr >= 19 ? 'text-green-500' : 'text-yellow-600';
-                        const subtitle = hr >= 22 ? "Get some rest so you're ready for tomorrow!" : hr >= 19 ? 'Enjoy the evening!' : 'Time to relax or tackle more tasks?';
-                        return (
-                          <div key="desktop-now-marker-end" className="flex gap-2.5 py-2.5">
-                            <div className={`w-1.5 rounded-full flex-shrink-0 ${barColor}`} />
-                            <div className="min-w-0 flex-1">
-                              <div className={`text-sm font-medium ${textColor}`}>{formatTime(agendaNowMarker.nowTimeStr)}, all done!</div>
-                              <div className={`text-xs italic ${textColor} mt-0.5`}>{subtitle}</div>
-                            </div>
-                          </div>
-                        );
-                      })()}
-                    </div>
-                  ); })()}
-
-                  {/* GLANCEahead — tomorrow preview */}
-                  {(() => {
-                    const isDayDone = (todayAgenda.length > 0 && !agendaNowMarker.insideTask && agendaNowMarker.insertAfterIndex >= todayAgenda.length - 1) || todayAgenda.length === 0;
-                    const isEvening = currentTime.getHours() >= 19;
-                    if (!isDayDone && !isEvening) return null;
-                    const { dayLabel, taskCount, eventCount, deadlineCount, firstStartTime, committedMinutes, isEmpty } = glanceAhead;
-                    const committedH = Math.floor(committedMinutes / 60);
-                    const committedM = committedMinutes % 60;
-                    const committedStr = committedH > 0 ? `${committedH}h${committedM > 0 ? ` ${committedM}m` : ''}` : committedM > 0 ? `${committedM}m` : null;
-                    return (
-                      <div className={`rounded-lg border ${borderClass} p-3`}>
-                        <div className={`text-xs font-semibold uppercase tracking-wide mb-2 ${textSecondary}`}>
-                          <span className="flex items-center gap-1.5">
-                            <span><span className="italic">GLANCE</span><span className="normal-case not-italic">ahead</span></span>
-                            <span className="font-normal normal-case">— {dayLabel}</span>
-                          </span>
-                        </div>
-                        {isEmpty ? (
-                          <p className={`text-sm ${textSecondary} italic`}>Tomorrow is wide open</p>
-                        ) : (
-                          <div className="space-y-1">
-                            {firstStartTime && (
-                              <div className="flex items-center gap-2">
-                                <Clock size={13} className={textSecondary} />
-                                <span className={`text-sm ${textPrimary}`}>Day starts at <span className="font-medium">{formatTime(firstStartTime)}</span></span>
-                              </div>
-                            )}
-                            <div className="flex flex-wrap items-center gap-x-3 gap-y-1">
-                              {taskCount > 0 && (
-                                <span className={`text-sm ${textPrimary} flex items-center gap-1`}><CheckSquare size={12} className={textSecondary} />{taskCount} task{taskCount !== 1 ? 's' : ''}</span>
-                              )}
-                              {eventCount > 0 && (
-                                <span className={`text-sm ${textPrimary} flex items-center gap-1`}><Calendar size={12} className={textSecondary} />{eventCount} event{eventCount !== 1 ? 's' : ''}</span>
-                              )}
-                              {deadlineCount > 0 && (
-                                <span className={`text-sm font-medium ${darkMode ? 'text-orange-400' : 'text-orange-600'} flex items-center gap-1`}><AlertTriangle size={12} />{deadlineCount} deadline{deadlineCount !== 1 ? 's' : ''}</span>
-                              )}
-                              {committedStr && (
-                                <span className={`text-xs px-2 py-0.5 rounded-full ${darkMode ? 'bg-blue-900/40 text-blue-300' : 'bg-blue-100 text-blue-700'}`}>{committedStr} committed</span>
-                              )}
-                            </div>
-                          </div>
-                        )}
-                      </div>
-                    );
-                  })()}
-
-                  {/* Routines row */}
-                  {routinesEnabled && todayRoutines.length > 0 && (() => {
-                    const nowMin = currentTime.getHours() * 60 + currentTime.getMinutes();
-                    const visibleRoutines = todayRoutines.filter(r => {
-                      if (String(r.id).startsWith('example-')) return false;
-                      if (!r.startTime || r.isAllDay) return true;
-                      return (timeToMinutes(r.startTime) + r.duration + 60) > nowMin;
-                    });
-                    if (visibleRoutines.length === 0) return null;
-                    return (
-                      <div className={`rounded-lg border ${borderClass} p-3 cursor-pointer hover:opacity-80 transition-opacity`} onClick={() => openRoutinesDashboard()}>
-                        <div className={`text-xs font-semibold uppercase tracking-wide mb-2 ${textSecondary}`}>Routines</div>
-                        <div className="flex flex-wrap gap-1">
-                          {[...visibleRoutines].sort((a, b) => {
-                            if (a.isAllDay && !b.isAllDay) return -1;
-                            if (!a.isAllDay && b.isAllDay) return 1;
-                            if (a.startTime && b.startTime) return timeToMinutes(a.startTime) - timeToMinutes(b.startTime);
-                            return 0;
-                          }).map(r => {
-                            let timeLabel = '';
-                            if (!r.isAllDay && r.startTime) {
-                              if (use24HourClock) {
-                                timeLabel = r.startTime;
-                              } else {
-                                const [h, m] = r.startTime.split(':').map(Number);
-                                const hour12 = h === 0 ? 12 : h > 12 ? h - 12 : h;
-                                const ampm = h < 12 ? 'a' : 'p';
-                                timeLabel = m === 0 ? `${hour12}${ampm}` : `${hour12}:${String(m).padStart(2, '0')}${ampm}`;
-                              }
-                            }
-                            return (
-                              <span key={r.id} className={`rounded-full px-2.5 py-0.5 text-xs font-medium ${darkMode ? 'bg-teal-700/80 text-teal-100' : 'bg-teal-600/80 text-white'}`}>
-                                {timeLabel && <span className="opacity-70 mr-1">{timeLabel}</span>}{r.name}
-                              </span>
-                            );
-                          })}
-                        </div>
-                      </div>
-                    );
-                  })()}
-
-                </div>
+                <GlanceSidebar variant="desktop" />
               </div>
-
               )}
 
               {/* Inbox section — shown when inbox tab active */}
